refactor!(cdt): split triangulation modules and harden invariants
- Move CDT triangulation builders, foliation logic, mutation hooks, validation, and simulation result summaries into focused modules.
- Validate Metropolis target and proposal parameters before constructing reusable MCMC components.
- Reject non-finite geometry generator inputs and invalid toroidal domains before they reach Delaunay predicates.
- Make Delaunay backend mutations restore prior state on upstream errors or malformed flip outputs.
- Return concrete topology iterators from TriangulationQuery to avoid boxed iterator dispatch in repeated scans.
- Add validated example output markers and production Rust Semgrep rules for panic-prone error paths.

BREAKING CHANGE: CdtTarget::new, CdtProposal::new, and CdtProposal::with_seed now return CdtResult, and TriangulationQuery iterator methods now return concrete impl Iterator values instead of boxed trait objects.
