+++ docs/dev/testing.md
@@ -199,12 +199,12 @@ Before proposing patches agents should run:
 just ci
 ```
 
-The `ci` recipe runs `check bench-compile test-all examples`, which enforces:
+The `ci` recipe runs `check bench-compile test-all validate-examples`, which enforces:
 
 - **check** (via `lint`): formatting, clippy, documentation builds, markdown, spelling, config validation (JSON, TOML, YAML, GitHub Actions)
 - **bench-compile**: benchmarks compile without warnings under `-D warnings`
 - **test-all**: unit tests, doc tests, integration tests, and Python tests (pytest)
-- **examples**: all example scripts run successfully
+- **validate-examples**: all Cargo examples run successfully and user-facing examples emit stable output markers
 
 ---
