+++ src/cdt/triangulation.rs
@@ -5,24 +5,20 @@
 //! This module provides CDT-specific triangulation data structures that work
 //! with any geometry backend implementing the trait interfaces.
 
-use crate::cdt::foliation::{CellType, EdgeType, Foliation, FoliationError, classify_cell};
+use crate::cdt::foliation::Foliation;
 use crate::config::CdtTopology;
 use crate::errors::{CdtError, CdtResult};
+#[cfg(test)]
 use crate::geometry::DelaunayBackend2D;
-use crate::geometry::backends::delaunay::{
-    DelaunayEdgeHandle, DelaunayError, DelaunayFaceHandle, DelaunayVertexHandle,
-};
 #[cfg(test)]
 use crate::geometry::generators::build_delaunay2_with_data;
-use crate::geometry::generators::{
-    build_delaunay2_from_cells, build_toroidal_delaunay2, generate_delaunay2,
-};
-use crate::geometry::traits::{
-    FlipResult, SubdivisionResult, TriangulationMut, TriangulationQuery,
-};
-use crate::util::f64_band_to_u32;
-use std::collections::{HashMap, HashSet};
+use crate::geometry::traits::TriangulationQuery;
 use std::time::Instant;
+
+mod builders;
+mod foliation;
+mod moves;
+mod validation;
 
 /// CDT-specific triangulation wrapper - completely geometry-agnostic
 #[derive(Debug, Clone)]
@@ -67,136 +63,6 @@ struct GeometryCache {
 struct CachedValue<T> {
     value: T,
     modification_count: u64,
-}
-
-/// Rewrites explicit toroidal builder failures with CDT-level generation context.
-///
-/// The lower geometry builder reports failures in terms of its input shape; this
-/// helper preserves the underlying diagnostic while normalizing the public error
-/// fields to the toroidal CDT constructor's vertex count, domain, and first attempt.
-fn remap_toroidal_generation_error(error: CdtError, total_vertices: u32) -> CdtError {
-    match error {
-        CdtError::DelaunayGenerationFailed {
-            underlying_error, ..
-        } => CdtError::DelaunayGenerationFailed {
-            vertex_count: total_vertices,
-            coordinate_range: (0.0, 1.0),
-            attempt: 1,
-            underlying_error,
-        },
-        other => other,
-    }
-}
-
-/// Rewrites explicit strip builder failures with CDT-level generation context.
-fn remap_strip_generation_error(
-    error: CdtError,
-    total_vertices: u32,
-    coordinate_max: f64,
-) -> CdtError {
-    match error {
-        CdtError::DelaunayGenerationFailed {
-            underlying_error, ..
-        } => CdtError::DelaunayGenerationFailed {
-            vertex_count: total_vertices,
-            coordinate_range: (0.0, coordinate_max),
-            attempt: 1,
-            underlying_error,
-        },
-        other => other,
-    }
-}
-
-/// Builds a CDT-level generation error for explicit strip construction failures.
-const fn strip_generation_error(
-    total_vertices: u32,
-    coordinate_max: f64,
-    underlying_error: String,
-) -> CdtError {
-    CdtError::DelaunayGenerationFailed {
-        vertex_count: total_vertices,
-        coordinate_range: (0.0, coordinate_max),
-        attempt: 1,
-        underlying_error,
-    }
-}
-
-/// Verifies that the explicit strip builder returned the requested mesh size.
-#[expect(
-    clippy::too_many_arguments,
-    reason = "count mismatch diagnostics preserve both requested CDT parameters and expected builder counts"
-)]
-fn validate_strip_counts(
-    backend: &DelaunayBackend2D,
-    total_vertices: u32,
-    total_cells: u32,
-    expected_vertices: usize,
-    expected_faces: usize,
-    vertices_per_slice: u32,
-    num_slices: u32,
-    coordinate_max: f64,
-) -> CdtResult<()> {
-    if backend.vertex_count() != expected_vertices {
-        return Err(strip_generation_error(
-            total_vertices,
-            coordinate_max,
-            format!(
-                "build_delaunay2_from_cells()/from_cdt_strip() produced {} vertices, expected {} for vertices_per_slice={} and num_slices={}",
-                backend.vertex_count(),
-                total_vertices,
-                vertices_per_slice,
-                num_slices,
-            ),
-        ));
-    }
-    if backend.face_count() != expected_faces {
-        return Err(strip_generation_error(
-            total_vertices,
-            coordinate_max,
-            format!(
-                "build_delaunay2_from_cells()/from_cdt_strip() produced {} faces, expected {} for vertices_per_slice={} and num_slices={}",
-                backend.face_count(),
-                total_cells,
-                vertices_per_slice,
-                num_slices,
-            ),
-        ));
-    }
-
-    Ok(())
-}
-
-/// Builds a CDT-level generation error for explicit toroidal construction failures.
-const fn toroidal_generation_error(total_vertices: u32, underlying_error: String) -> CdtError {
-    CdtError::DelaunayGenerationFailed {
-        vertex_count: total_vertices,
-        coordinate_range: (0.0, 1.0),
-        attempt: 1,
-        underlying_error,
-    }
-}
-
-/// Verifies that the explicit toroidal builder returned the requested mesh size.
-fn validate_toroidal_counts(
-    backend: &DelaunayBackend2D,
-    total_vertices: u32,
-    expected_vertices: usize,
-    expected_faces: usize,
-) -> CdtResult<()> {
-    if backend.vertex_count() != expected_vertices || backend.face_count() != expected_faces {
-        return Err(toroidal_generation_error(
-            total_vertices,
-            format!(
-                "explicit toroidal builder produced {} vertices and {} faces, expected {} vertices and {} faces",
-                backend.vertex_count(),
-                backend.face_count(),
-                total_vertices,
-                expected_faces,
-            ),
-        ));
-    }
-
-    Ok(())
 }
 
 /// Events in simulation history
@@ -719,1898 +585,6 @@ impl<B: TriangulationQuery> CdtTriangulation<B> {
     }
 }
 
-impl CdtTriangulation<DelaunayBackend2D> {
-    /// Validate foliation consistency.
-    ///
-    /// If no foliation is present, succeeds vacuously.
-    /// Otherwise checks:
-    /// 1. The stored labeled-vertex count matches the geometry vertex count
-    /// 2. Every stored time slice is non-empty
-    /// 3. Live backend labels match stored per-slice bookkeeping
-    ///
-    /// # Errors
-    ///
-    /// Returns error if foliation structure is invalid.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::prelude::triangulation::*;
-    ///
-    /// let mut tri = CdtTriangulation::from_seeded_points(12, 3, 2, 42)
-    ///     .expect("create seeded triangulation");
-    /// tri.assign_foliation_by_y(3)
-    ///     .expect("assign foliation from y-coordinates");
-    /// assert!(tri.validate_foliation().is_ok());
-    /// ```
-    pub fn validate_foliation(&self) -> CdtResult<()> {
-        let Some(foliation) = &self.foliation else {
-            return Ok(());
-        };
-
-        // Check that all vertices are labeled
-        let vertex_count = self.geometry.vertex_count();
-        if foliation.labeled_vertex_count() != vertex_count {
-            return Err(FoliationError::LabelCountMismatch {
-                labeled: foliation.labeled_vertex_count(),
-                expected: vertex_count,
-            }
-            .into());
-        }
-
-        // Check that every slice is non-empty
-        for (t, &size) in foliation.slice_sizes().iter().enumerate() {
-            if size == 0 {
-                return Err(FoliationError::EmptySlice { slice: t }.into());
-            }
-        }
-
-        // Validate against live labels from canonical Delaunay vertex payload.
-        let mut live_slice_sizes = vec![0usize; foliation.slice_sizes().len()];
-
-        for (vertex, vh) in self.geometry.vertices().enumerate() {
-            let Some(label) = self.geometry.vertex_data_by_key(vh.vertex_key()) else {
-                return Err(FoliationError::MissingVertexLabel { vertex }.into());
-            };
-
-            let slice = label as usize;
-            if slice >= live_slice_sizes.len() {
-                return Err(FoliationError::OutOfRangeVertexLabel {
-                    vertex,
-                    label,
-                    expected_range_end: live_slice_sizes.len(),
-                }
-                .into());
-            }
-
-            live_slice_sizes[slice] += 1;
-        }
-
-        for (slice, (&expected, &actual)) in foliation
-            .slice_sizes()
-            .iter()
-            .zip(live_slice_sizes.iter())
-            .enumerate()
-        {
-            if expected != actual {
-                return Err(FoliationError::LabelMismatch {
-                    slice,
-                    expected,
-                    actual,
-                }
-                .into());
-            }
-        }
-
-        // Toroidal topology adds two stronger structural invariants:
-        //  - every spatial slice forms a closed S¹ (each vertex has exactly
-        //    two spacelike neighbours and the subgraph of each slice is a
-        //    single cycle); and
-        //  - the time direction wraps: every slice has timelike adjacency to
-        //    both `(t-1) mod T` and `(t+1) mod T`.  χ = 0 alone does not
-        //    distinguish a torus from a cylinder, so we check the wrap
-        //    explicitly.
-        if matches!(self.metadata.topology, CdtTopology::Toroidal) {
-            self.validate_toroidal_spatial_rings()?;
-            self.validate_toroidal_temporal_wraparound()?;
-        }
-
-        Ok(())
-    }
-
-    /// Validates that the time direction wraps for toroidal topology.
-    ///
-    /// For each slice `t` (with `T = time_slices`), checks that the foliation
-    /// has at least one timelike edge to slice `(t - 1) mod T` and at least
-    /// one to slice `(t + 1) mod T`.  Cylinders (open in time) have χ = 0
-    /// like the torus, so this check is what actually distinguishes them.
-    fn validate_toroidal_temporal_wraparound(&self) -> CdtResult<()> {
-        let total = self.metadata.time_slices;
-        if total < 2 {
-            return Ok(());
-        }
-        let num_slices = total as usize;
-
-        // Collect, per slice, the set of slices reached via timelike (step
-        // distance == 1) edges.
-        let mut neighbor_slices: Vec<HashSet<usize>> = vec![HashSet::new(); num_slices];
-        for edge in self.geometry.edges() {
-            let Some((v0, v1)) = self.geometry.edge_endpoints(&edge) else {
-                continue;
-            };
-            let Some(t0) = self.geometry.vertex_data_by_key(v0.vertex_key()) else {
-                continue;
-            };
-            let Some(t1) = self.geometry.vertex_data_by_key(v1.vertex_key()) else {
-                continue;
-            };
-            if t0 >= total || t1 >= total {
-                continue;
-            }
-            if self.time_step_distance(t0, t1) != 1 {
-                continue;
-            }
-            let s0 = t0 as usize;
-            let s1 = t1 as usize;
-            neighbor_slices[s0].insert(s1);
-            neighbor_slices[s1].insert(s0);
-        }
-
-        for (slice, neighbors) in neighbor_slices.iter().enumerate() {
-            let prev = if slice == 0 {
-                num_slices - 1
-            } else {
-                slice - 1
-            };
-            let next = (slice + 1) % num_slices;
-            if !neighbors.contains(&prev) {
-                return Err(FoliationError::MissingTemporalWrapAround {
-                    slice,
-                    missing_neighbor: prev,
-                }
-                .into());
-            }
-            if !neighbors.contains(&next) {
-                return Err(FoliationError::MissingTemporalWrapAround {
-                    slice,
-                    missing_neighbor: next,
-                }
-                .into());
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Validates that every spatial slice forms a closed S¹.
-    ///
-    /// For each time slice `t` we iterate over backend edges, count incident
-    /// spacelike edges (`|Δt| = 0`) per vertex, and walk the resulting
-    /// spacelike subgraph to verify it is a single cycle of length
-    /// `slice_sizes[t]`.
-    fn validate_toroidal_spatial_rings(&self) -> CdtResult<()> {
-        let Some(foliation) = &self.foliation else {
-            return Ok(());
-        };
-
-        let num_slices = foliation.slice_sizes().len();
-        let mut spacelike_neighbors: Vec<HashMap<DelaunayVertexHandle, Vec<DelaunayVertexHandle>>> =
-            vec![HashMap::new(); num_slices];
-
-        for edge in self.geometry.edges() {
-            let Some((v0, v1)) = self.geometry.edge_endpoints(&edge) else {
-                continue;
-            };
-            let Some(t0) = self.geometry.vertex_data_by_key(v0.vertex_key()) else {
-                continue;
-            };
-            let Some(t1) = self.geometry.vertex_data_by_key(v1.vertex_key()) else {
-                continue;
-            };
-            if t0 != t1 {
-                continue;
-            }
-            let slice = t0 as usize;
-            if slice >= num_slices {
-                continue;
-            }
-            spacelike_neighbors[slice]
-                .entry(v0.clone())
-                .or_default()
-                .push(v1.clone());
-            spacelike_neighbors[slice].entry(v1).or_default().push(v0);
-        }
-
-        for (slice, adjacency) in spacelike_neighbors.iter().enumerate() {
-            let expected_size = foliation.slice_sizes()[slice];
-            if adjacency.len() != expected_size {
-                return Err(FoliationError::SpacelikeSubgraphSizeMismatch {
-                    slice,
-                    observed: adjacency.len(),
-                    expected: expected_size,
-                }
-                .into());
-            }
-            for (vertex, neighbors) in adjacency {
-                if neighbors.len() != 2 {
-                    return Err(FoliationError::SpacelikeDegreeViolation {
-                        slice,
-                        vertex: format!("{:?}", vertex.vertex_key()),
-                        observed_degree: neighbors.len(),
-                    }
-                    .into());
-                }
-            }
-
-            // Walk the cycle starting from any vertex; verify it visits every
-            // vertex of the slice and closes back on itself.
-            let Some(start) = adjacency.keys().next() else {
-                continue;
-            };
-            let mut visited: HashSet<DelaunayVertexHandle> = HashSet::new();
-            visited.insert(start.clone());
-            let mut prev = start.clone();
-            let mut current = adjacency[start][0].clone();
-            while current != *start {
-                if !visited.insert(current.clone()) {
-                    // Non-simple cycle: a vertex was revisited before the walk
-                    // returned to `start`.  Surface this through the same
-                    // typed variant as a short cycle — both indicate the
-                    // spacelike subgraph isn't a single closed S¹.
-                    return Err(FoliationError::SpacelikeNonClosedRing {
-                        slice,
-                        walked: visited.len(),
-                        expected: expected_size,
-                    }
-                    .into());
-                }
-                let neighbors = &adjacency[&current];
-                let next = if neighbors[0] == prev {
-                    neighbors[1].clone()
-                } else {
-                    neighbors[0].clone()
-                };
-                prev = current;
-                current = next;
-            }
-            if visited.len() != expected_size {
-                return Err(FoliationError::SpacelikeNonClosedRing {
-                    slice,
-                    walked: visited.len(),
-                    expected: expected_size,
-                }
-                .into());
-            }
-        }
-
-        Ok(())
-    }
-}
-
-// =============================================================================
-// Delaunay-specific factory functions and foliation methods
-// =============================================================================
-impl CdtTriangulation<DelaunayBackend2D> {
-    /// Flips an edge and marks CDT-derived state stale when the backend mutation succeeds.
-    pub(crate) fn flip_edge(
-        &mut self,
-        edge: DelaunayEdgeHandle,
-    ) -> Result<FlipResult<DelaunayEdgeHandle, DelaunayFaceHandle>, DelaunayError> {
-        let result = self.geometry.flip_edge(edge)?;
-        self.bump_modification_count();
-        Ok(result)
-    }
-
-    /// Subdivides a face and marks CDT-derived state stale when the backend mutation succeeds.
-    pub(crate) fn subdivide_face(
-        &mut self,
-        face: DelaunayFaceHandle,
-        point: &[f64],
-    ) -> Result<SubdivisionResult<DelaunayVertexHandle, DelaunayFaceHandle>, DelaunayError> {
-        let result = self.geometry.subdivide_face(face, point)?;
-        self.bump_modification_count();
-        Ok(result)
-    }
-
-    /// Removes a vertex and marks CDT-derived state stale when the backend mutation succeeds.
-    pub(crate) fn remove_vertex(
-        &mut self,
-        vertex: DelaunayVertexHandle,
-    ) -> Result<Vec<DelaunayFaceHandle>, DelaunayError> {
-        let result = self.geometry.remove_vertex(vertex)?;
-        self.bump_modification_count();
-        Ok(result)
-    }
-
-    /// Updates a vertex time label and marks CDT-derived state stale on success.
-    pub(crate) fn set_vertex_data(
-        &mut self,
-        vertex: &DelaunayVertexHandle,
-        data: Option<u32>,
-    ) -> Result<(), DelaunayError> {
-        self.geometry
-            .set_vertex_data_by_key(vertex.vertex_key(), data)?;
-        self.bump_modification_count();
-        Ok(())
-    }
-
-    // -------------------------------------------------------------------------
-    // Factory constructors
-    // -------------------------------------------------------------------------
-
-    /// Re-reads current backend labels during validation so stale stored bookkeeping is detected.
-    fn live_slice_sizes_from_vertex_labels(
-        backend: &DelaunayBackend2D,
-        num_slices: u32,
-    ) -> CdtResult<Vec<usize>> {
-        if num_slices == 0 {
-            return Err(FoliationError::SliceSizeMismatch {
-                slice_sizes_len: 0,
-                num_slices,
-            }
-            .into());
-        }
-
-        let mut slice_sizes = vec![0usize; num_slices as usize];
-
-        for (vertex, vh) in backend.vertices().enumerate() {
-            if let Some(t) = backend.vertex_data_by_key(vh.vertex_key()) {
-                let idx = t as usize;
-                if idx >= slice_sizes.len() {
-                    return Err(FoliationError::OutOfRangeVertexLabel {
-                        vertex,
-                        label: t,
-                        expected_range_end: slice_sizes.len(),
-                    }
-                    .into());
-                }
-                slice_sizes[idx] += 1;
-            } else {
-                return Err(FoliationError::MissingVertexLabel { vertex }.into());
-            }
-        }
-
-        Ok(slice_sizes)
-    }
-
-    /// Create a new CDT triangulation with Delaunay backend from random points.
-    ///
-    /// This is the recommended way to create triangulations for simulations.
-    ///
-    /// # Errors
-    /// Returns error if triangulation generation fails
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::CdtTriangulation;
-    ///
-    /// let tri = CdtTriangulation::from_random_points(5, 2, 2).unwrap();
-    /// assert_eq!(tri.time_slices(), 2);
-    /// ```
-    pub fn from_random_points(vertices: u32, time_slices: u32, dimension: u8) -> CdtResult<Self> {
-        // Validate dimension first
-        if dimension != 2 {
-            return Err(CdtError::UnsupportedDimension(dimension.into()));
-        }
-
-        // Validate other parameters
-        if vertices < 3 {
-            return Err(CdtError::InvalidGenerationParameters {
-                issue: "Insufficient vertex count".to_string(),
-                provided_value: vertices.to_string(),
-                expected_range: "≥ 3".to_string(),
-            });
-        }
-
-        let dt = generate_delaunay2(vertices, (0.0, 10.0), None)?;
-        let backend = DelaunayBackend2D::from_triangulation(dt);
-
-        Self::try_new(backend, time_slices, dimension)
-    }
-
-    /// Create a new CDT triangulation with Delaunay backend from random points using a fixed seed.
-    ///
-    /// This function provides deterministic triangulation generation for testing purposes.
-    ///
-    /// # Errors
-    /// Returns error if triangulation generation fails
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::CdtTriangulation;
-    ///
-    /// let tri = CdtTriangulation::from_seeded_points(5, 2, 2, 53).unwrap();
-    /// assert_eq!(tri.vertex_count(), 5);
-    /// ```
-    pub fn from_seeded_points(
-        vertices: u32,
-        time_slices: u32,
-        dimension: u8,
-        seed: u64,
-    ) -> CdtResult<Self> {
-        // Validate dimension first
-        if dimension != 2 {
-            return Err(CdtError::UnsupportedDimension(dimension.into()));
-        }
-
-        // Validate other parameters
-        if vertices < 3 {
-            return Err(CdtError::InvalidGenerationParameters {
-                issue: "Insufficient vertex count".to_string(),
-                provided_value: vertices.to_string(),
-                expected_range: "≥ 3".to_string(),
-            });
-        }
-
-        let dt = generate_delaunay2(vertices, (0.0, 10.0), Some(seed))?;
-        let backend = DelaunayBackend2D::from_triangulation(dt);
-
-        Self::try_new(backend, time_slices, dimension)
-    }
-
-    /// Wrap a labeled 2D Delaunay backend and derive foliation from vertex data.
-    ///
-    /// Preserves per-vertex time labels already embedded in the backend.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`CdtError::UnsupportedDimension`] if `dimension != 2`.
-    /// Returns [`CdtError::ValidationFailed`] if any vertex is unlabeled or
-    /// has a time label outside `0..time_slices`, or if any time slice is empty.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::prelude::geometry::*;
-    /// use causal_triangulations::prelude::triangulation::*;
-    ///
-    /// let dt = build_delaunay2_with_data(&[
-    ///     ([0.0, 0.0], 0),
-    ///     ([1.0, 0.0], 0),
-    ///     ([0.5, 1.0], 1),
-    /// ])
-    /// .expect("build labeled triangle");
-    /// let backend = DelaunayBackend2D::from_triangulation(dt);
-    /// let tri = CdtTriangulation::from_labeled_delaunay(backend, 2, 2)
-    ///     .expect("wrap labeled backend");
-    ///
-    /// assert!(tri.has_foliation());
-    /// assert_eq!(tri.slice_sizes(), &[2, 1]);
-    /// ```
-    pub fn from_labeled_delaunay(
-        backend: DelaunayBackend2D,
-        time_slices: u32,
-        dimension: u8,
-    ) -> CdtResult<Self> {
-        if dimension != 2 {
-            return Err(CdtError::UnsupportedDimension(dimension.into()));
-        }
-
-        Self::check_time_slices(CdtTopology::OpenBoundary, time_slices)?;
-        let slice_sizes = Self::live_slice_sizes_from_vertex_labels(&backend, time_slices)?;
-        for (slice, &size) in slice_sizes.iter().enumerate() {
-            if size == 0 {
-                return Err(FoliationError::EmptySlice { slice }.into());
-            }
-        }
-        let foliation =
-            Foliation::from_slice_sizes(slice_sizes, time_slices).map_err(CdtError::from)?;
-
-        let mut tri = Self::try_new(backend, time_slices, dimension)?;
-        tri.foliation = Some(foliation);
-        tri.mark_foliation_synchronized();
-        Ok(tri)
-    }
-
-    /// Construct a true 1+1 CDT strip by explicit layered connectivity.
-    ///
-    /// Places `vertices_per_slice` vertices on each open spatial slice and
-    /// connects adjacent time slices into quads. Each quad is split into one
-    /// Up `(2,1)` triangle and one Down `(1,2)` triangle, so every finite face
-    /// is classifiable by construction.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`CdtError::InvalidGenerationParameters`] if `vertices_per_slice < 4`,
-    /// `num_slices < 2`, or the derived vertex or cell count overflows `u32`.
-    /// Returns [`CdtError::DelaunayGenerationFailed`] if constructor storage cannot
-    /// be reserved, if the underlying explicit builder rejects the mesh, or if
-    /// `build_delaunay2_from_cells()` returns a vertex or face count that does not
-    /// match the requested strip. Returns [`CdtError::Foliation`],
-    /// [`CdtError::CausalityViolation`], or [`CdtError::ValidationFailed`] if the
-    /// constructed strip fails CDT validation.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::prelude::triangulation::*;
-    ///
-    /// let tri = CdtTriangulation::from_cdt_strip(4, 2)
-    ///     .expect("build explicit CDT strip");
-    /// assert_eq!(tri.vertex_count(), 8);
-    /// assert_eq!(tri.face_count(), 6);
-    /// assert!(tri.validate_cell_classification().is_ok());
-    /// ```
-    #[expect(
-        clippy::too_many_lines,
-        reason = "explicit strip construction includes fallible allocation handling and post-build validation"
-    )]
-    pub fn from_cdt_strip(vertices_per_slice: u32, num_slices: u32) -> CdtResult<Self> {
-        if vertices_per_slice < 4 {
-            return Err(CdtError::InvalidGenerationParameters {
-                issue: "Insufficient vertices per slice".to_string(),
-                provided_value: vertices_per_slice.to_string(),
-                expected_range: "≥ 4".to_string(),
-            });
-        }
-        if num_slices < 2 {
-            return Err(CdtError::InvalidGenerationParameters {
-                issue: "Insufficient number of time slices".to_string(),
-                provided_value: num_slices.to_string(),
-                expected_range: "≥ 2".to_string(),
-            });
-        }
-
-        let total_vertices = vertices_per_slice.checked_mul(num_slices).ok_or_else(|| {
-            CdtError::InvalidGenerationParameters {
-                issue: "Vertex count overflow".to_string(),
-                provided_value: format!("{vertices_per_slice} × {num_slices}"),
-                expected_range: "product ≤ u32::MAX".to_string(),
-            }
-        })?;
-
-        let spatial_quads = vertices_per_slice - 1;
-        let temporal_quads = num_slices - 1;
-        let total_quads = spatial_quads.checked_mul(temporal_quads).ok_or_else(|| {
-            CdtError::InvalidGenerationParameters {
-                issue: "Cell count overflow".to_string(),
-                provided_value: format!("{spatial_quads} × {temporal_quads}"),
-                expected_range: "product ≤ u32::MAX".to_string(),
-            }
-        })?;
-        let total_cells =
-            total_quads
-                .checked_mul(2)
-                .ok_or_else(|| CdtError::InvalidGenerationParameters {
-                    issue: "Cell count overflow".to_string(),
-                    provided_value: format!("2 × {total_quads}"),
-                    expected_range: "product ≤ u32::MAX".to_string(),
-                })?;
-
-        let coordinate_max = f64::from(num_slices - 1).max(1.0);
-        let generation_failed = |underlying_error: String| {
-            strip_generation_error(total_vertices, coordinate_max, underlying_error)
-        };
-
-        let expected_vertices =
-            usize::try_from(total_vertices).map_err(|err| generation_failed(err.to_string()))?;
-        let expected_faces =
-            usize::try_from(total_cells).map_err(|err| generation_failed(err.to_string()))?;
-
-        let n = usize::try_from(vertices_per_slice)
-            .map_err(|err| generation_failed(err.to_string()))?;
-        let t_count =
-            usize::try_from(num_slices).map_err(|err| generation_failed(err.to_string()))?;
-        let index = |i: usize, t: usize| -> usize { t * n + i };
-
-        let spacing = 1.0_f64 / f64::from(vertices_per_slice - 1);
-        let mut vertex_specs: Vec<([f64; 2], u32)> = Vec::new();
-        vertex_specs
-            .try_reserve_exact(expected_vertices)
-            .map_err(|err| {
-                generation_failed(format!(
-                    "from_cdt_strip() failed to reserve {expected_vertices} vertex specs for vertices_per_slice={vertices_per_slice}, num_slices={num_slices}: {err}"
-                ))
-            })?;
-        for t in 0..num_slices {
-            for i in 0..vertices_per_slice {
-                vertex_specs.push(([f64::from(i) * spacing, f64::from(t)], t));
-            }
-        }
-
-        let mut cells: Vec<[usize; 3]> = Vec::new();
-        cells.try_reserve_exact(expected_faces).map_err(|err| {
-            generation_failed(format!(
-                "from_cdt_strip() failed to reserve {expected_faces} triangle cells for vertices_per_slice={vertices_per_slice}, num_slices={num_slices}: {err}"
-            ))
-        })?;
-        for t in 0..(t_count - 1) {
-            let t_next = t + 1;
-            for i in 0..(n - 1) {
-                let i_next = i + 1;
-                cells.push([index(i, t), index(i_next, t), index(i, t_next)]);
-                cells.push([index(i_next, t), index(i_next, t_next), index(i, t_next)]);
-            }
-        }
-
-        // delaunay 0.7.6 accepts explicit cells as Vec-backed index lists.
-        // Keep the strip working set compact, then adapt fallibly at the API boundary.
-        let mut cell_specs: Vec<Vec<usize>> = Vec::new();
-        cell_specs.try_reserve_exact(expected_faces).map_err(|err| {
-            generation_failed(format!(
-                "from_cdt_strip() failed to reserve {expected_faces} builder cell specs for build_delaunay2_from_cells(): {err}"
-            ))
-        })?;
-        for cell in &cells {
-            let mut cell_spec = Vec::new();
-            cell_spec.try_reserve_exact(3).map_err(|err| {
-                generation_failed(format!(
-                    "from_cdt_strip() failed to reserve a build_delaunay2_from_cells() triangle cell spec: {err}"
-                ))
-            })?;
-            cell_spec.extend_from_slice(cell);
-            cell_specs.push(cell_spec);
-        }
-
-        let dt = build_delaunay2_from_cells(&vertex_specs, &cell_specs)
-            .map_err(|err| remap_strip_generation_error(err, total_vertices, coordinate_max))?;
-
-        let backend = DelaunayBackend2D::from_triangulation(dt);
-        validate_strip_counts(
-            &backend,
-            total_vertices,
-            total_cells,
-            expected_vertices,
-            expected_faces,
-            vertices_per_slice,
-            num_slices,
-            coordinate_max,
-        )?;
-
-        let slice_sizes = vec![n; t_count];
-        let foliation =
-            Foliation::from_slice_sizes(slice_sizes, num_slices).map_err(CdtError::from)?;
-
-        let mut tri = Self::try_new(backend, num_slices, 2)?;
-        tri.foliation = Some(foliation);
-        tri.mark_foliation_synchronized();
-
-        tri.validate_foliation()?;
-        tri.validate_causality_delaunay()?;
-        tri.validate_topology()?;
-        tri.classify_all_cells()?;
-
-        Ok(tri)
-    }
-
-    /// Construct a foliated 1+1 CDT on a torus (S¹×S¹).
-    ///
-    /// Places `vertices_per_slice` vertices per time slice, uniformly spaced
-    /// on S¹ (spatial coordinate periodic in `[0, 1)`).  Time slices wrap:
-    /// slice `num_slices - 1` connects back to slice `0`.  Each quad between
-    /// adjacent slices is split into one Up (2,1) and one Down (1,2) triangle.
-    ///
-    /// The triangulation is built by explicit combinatorial connectivity via
-    /// [`crate::geometry::generators::build_toroidal_delaunay2`],
-    /// which sets `TopologyGuarantee::Pseudomanifold` and
-    /// `GlobalTopology::Toroidal` so the underlying validator expects χ = 0.
-    ///
-    /// # Mesh structure
-    ///
-    /// With `N = vertices_per_slice` and `T = num_slices` the resulting mesh
-    /// has `N · T` vertices, `3 · N · T` edges, and `2 · N · T` triangles
-    /// (`V − E + F = 0`, the Euler characteristic of the torus).  Each pair of
-    /// adjacent slices `(t, t+1) mod T` and each spatial pair `(i, i+1) mod N`
-    /// contribute exactly one Up `(i, t), (i+1, t), (i, t+1)` and one Down
-    /// `(i+1, t), (i+1, t+1), (i, t+1)` triangle, so every triangle has
-    /// exactly one spacelike edge and two timelike edges by construction.
-    ///
-    /// # Arguments
-    ///
-    /// * `vertices_per_slice` — Number of vertices in each spatial slice (≥ 3).
-    /// * `num_slices` — Number of time slices (≥ 3 to keep `t-1` and `t+1`
-    ///   distinct after wrap-around).
-    ///
-    /// # Errors
-    ///
-    /// Returns [`CdtError::InvalidGenerationParameters`] if `vertices_per_slice < 3`
-    /// or `num_slices < 3`, or if the derived vertex or face count overflows `u32`.
-    /// Returns [`CdtError::DelaunayGenerationFailed`] if the underlying explicit
-    /// builder rejects the mesh, if constructor storage cannot be reserved, or if
-    /// the builder returns a vertex or face count that does not match the requested
-    /// toroidal CDT. Returns [`CdtError::Foliation`],
-    /// [`CdtError::CausalityViolation`], or [`CdtError::ValidationFailed`] if the
-    /// constructed triangulation fails CDT validation.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::prelude::triangulation::*;
-    ///
-    /// let tri = CdtTriangulation::from_toroidal_cdt(4, 3)
-    ///     .expect("build toroidal CDT");
-    /// assert_eq!(tri.vertex_count(), 12);
-    /// assert_eq!(tri.face_count(), 24);
-    /// assert!(tri.has_foliation());
-    /// ```
-    #[expect(
-        clippy::too_many_lines,
-        reason = "explicit toroidal construction includes fallible allocation handling and post-build validation"
-    )]
-    pub fn from_toroidal_cdt(vertices_per_slice: u32, num_slices: u32) -> CdtResult<Self> {
-        if vertices_per_slice < 3 {
-            return Err(CdtError::InvalidGenerationParameters {
-                issue: "Insufficient vertices per slice".to_string(),
-                provided_value: vertices_per_slice.to_string(),
-                expected_range: "≥ 3".to_string(),
-            });
-        }
-        if num_slices < 3 {
-            // With T=2 the wrap-around makes every pair of adjacent slices
-            // identify (t-1, t) with (t, t+1), so each spatial edge would be
-            // shared by 4 triangles instead of 2 — a non-manifold mesh.
-            return Err(CdtError::InvalidGenerationParameters {
-                issue: "Insufficient number of time slices".to_string(),
-                provided_value: num_slices.to_string(),
-                expected_range: "≥ 3".to_string(),
-            });
-        }
-
-        let total_vertices = vertices_per_slice.checked_mul(num_slices).ok_or_else(|| {
-            CdtError::InvalidGenerationParameters {
-                issue: "Vertex count overflow".to_string(),
-                provided_value: format!("{vertices_per_slice} × {num_slices}"),
-                expected_range: "product ≤ u32::MAX".to_string(),
-            }
-        })?;
-        let total_cells =
-            total_vertices
-                .checked_mul(2)
-                .ok_or_else(|| CdtError::InvalidGenerationParameters {
-                    issue: "Cell count overflow".to_string(),
-                    provided_value: format!("2 × {total_vertices}"),
-                    expected_range: "product ≤ u32::MAX".to_string(),
-                })?;
-
-        let generation_failed =
-            |underlying_error: String| toroidal_generation_error(total_vertices, underlying_error);
-
-        let expected_vertices =
-            usize::try_from(total_vertices).map_err(|err| generation_failed(err.to_string()))?;
-        let expected_faces =
-            usize::try_from(total_cells).map_err(|err| generation_failed(err.to_string()))?;
-
-        let n = usize::try_from(vertices_per_slice)
-            .map_err(|err| generation_failed(err.to_string()))?;
-        let t_count =
-            usize::try_from(num_slices).map_err(|err| generation_failed(err.to_string()))?;
-
-        // Index helper: vertex (i, t) → i + t * N. Both axes are periodic.
-        let index = |i: usize, t: usize| -> usize { (t % t_count) * n + (i % n) };
-
-        // --- Vertex coordinates (S¹ × S¹) ---
-        //
-        // Spatial coordinate: x_i = i / N is periodic in [0, 1).
-        // Time coordinate: t_t = t / T is periodic in [0, 1) so the metadata
-        // domain matches what we pass to GlobalTopology::Toroidal.
-        let n_f = f64::from(vertices_per_slice);
-        let t_f = f64::from(num_slices);
-        let mut vertex_specs: Vec<([f64; 2], u32)> = Vec::new();
-        vertex_specs
-            .try_reserve_exact(expected_vertices)
-            .map_err(|err| {
-                generation_failed(format!(
-                    "from_toroidal_cdt() failed to reserve {expected_vertices} vertex specs for vertices_per_slice={vertices_per_slice}, num_slices={num_slices}: {err}"
-                ))
-            })?;
-        for t in 0..num_slices {
-            for i in 0..vertices_per_slice {
-                let x = f64::from(i) / n_f;
-                let y = f64::from(t) / t_f;
-                vertex_specs.push(([x, y], t));
-            }
-        }
-
-        // --- Explicit cells (Up + Down per (i, t) quad) ---
-        let mut cells: Vec<[usize; 3]> = Vec::new();
-        cells.try_reserve_exact(expected_faces).map_err(|err| {
-            generation_failed(format!(
-                "from_toroidal_cdt() failed to reserve {expected_faces} triangle cells for vertices_per_slice={vertices_per_slice}, num_slices={num_slices}: {err}"
-            ))
-        })?;
-        for t in 0..t_count {
-            let t_next = (t + 1) % t_count;
-            for i in 0..n {
-                let i_next = (i + 1) % n;
-                // Up (2,1): two vertices on slice t, one on slice t+1.
-                cells.push([index(i, t), index(i_next, t), index(i, t_next)]);
-                // Down (1,2): one vertex on slice t, two on slice t+1.
-                cells.push([index(i_next, t), index(i_next, t_next), index(i, t_next)]);
-            }
-        }
-
-        // delaunay 0.7.6 accepts explicit cells as Vec-backed index lists.
-        // Keep the toroidal working set compact, then adapt fallibly at the API boundary.
-        let mut cell_specs: Vec<Vec<usize>> = Vec::new();
-        cell_specs.try_reserve_exact(expected_faces).map_err(|err| {
-            generation_failed(format!(
-                "from_toroidal_cdt() failed to reserve {expected_faces} builder cell specs for build_toroidal_delaunay2(): {err}"
-            ))
-        })?;
-        for cell in &cells {
-            let mut cell_spec = Vec::new();
-            cell_spec.try_reserve_exact(3).map_err(|err| {
-                generation_failed(format!(
-                    "from_toroidal_cdt() failed to reserve a build_toroidal_delaunay2() triangle cell spec: {err}"
-                ))
-            })?;
-            cell_spec.extend_from_slice(cell);
-            cell_specs.push(cell_spec);
-        }
-
-        let domain = [1.0_f64, 1.0_f64];
-        let dt = build_toroidal_delaunay2(&vertex_specs, &cell_specs, domain)
-            .map_err(|e| remap_toroidal_generation_error(e, total_vertices))?;
-
-        let backend = DelaunayBackend2D::from_triangulation(dt);
-        validate_toroidal_counts(&backend, total_vertices, expected_vertices, expected_faces)?;
-
-        let slice_sizes = vec![n; t_count];
-        let foliation =
-            Foliation::from_slice_sizes(slice_sizes, num_slices).map_err(CdtError::from)?;
-
-        let mut tri = Self::with_topology(backend, num_slices, 2, CdtTopology::Toroidal)?;
-        tri.foliation = Some(foliation);
-        tri.mark_foliation_synchronized();
-
-        // Propagate inner errors as-is so callers can pattern-match on the
-        // typed variant (e.g. `FoliationError::SpacelikeNonClosedRing` or
-        // `CausalityViolation`) instead of parsing a wrapped string.  Each
-        // inner validator already produces a precise, structured error.
-        tri.validate_foliation()?;
-        tri.validate_causality_delaunay()?;
-        tri.validate_topology()?;
-        tri.classify_all_cells()?;
-
-        Ok(tri)
-    }
-
-    // -------------------------------------------------------------------------
-    // Foliation assignment
-    // -------------------------------------------------------------------------
-
-    /// Assign a foliation to an existing triangulation by binning vertices
-    /// by their y-coordinate into `num_slices` equal bands.
-    ///
-    /// The y-coordinate range is determined from the actual vertex coordinates.
-    /// Band `t` covers `[y_min + t * band_height, y_min + (t+1) * band_height)`.
-    /// Time labels are written directly to vertex data.
-    ///
-    /// This is approximate — useful for testing but not guaranteed to produce
-    /// a valid causal structure.
-    ///
-    /// # Errors
-    ///
-    /// Returns error if `num_slices` is zero, if vertex coordinates cannot be
-    /// read, if y-bucket assignment would leave any time slice empty, if the
-    /// requested slice count violates the triangulation topology, or if writing
-    /// vertex labels or clearing stale cell labels in the backend fails.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::prelude::triangulation::*;
-    ///
-    /// let mut tri = CdtTriangulation::from_seeded_points(12, 3, 2, 42)
-    ///     .expect("create seeded triangulation");
-    /// tri.assign_foliation_by_y(3)
-    ///     .expect("assign foliation from y-coordinates");
-    ///
-    /// assert!(tri.has_foliation());
-    /// assert_eq!(tri.slice_sizes().iter().sum::<usize>(), tri.vertex_count());
-    /// ```
-    #[expect(
-        clippy::too_many_lines,
-        reason = "foliation assignment stages labels, writes backend payloads, and rolls back on failure to preserve atomic metadata/foliation invariants"
-    )]
-    pub fn assign_foliation_by_y(&mut self, num_slices: u32) -> CdtResult<()> {
-        if num_slices == 0 {
-            return Err(CdtError::InvalidGenerationParameters {
-                issue: "Number of slices must be positive".to_string(),
-                provided_value: "0".to_string(),
-                expected_range: "≥ 1".to_string(),
-            });
-        }
-
-        // Collect all vertex y-coordinates, failing fast if any vertex is unreadable.
-        let y_coords: Vec<(DelaunayVertexHandle, f64)> = self
-            .geometry
-            .vertices()
-            .map(|vh| {
-                let coords = self.geometry.vertex_coordinates(&vh).map_err(|e| {
-                    CdtError::ValidationFailed {
-                        check: "foliation_assignment".to_string(),
-                        detail: format!(
-                            "failed to read coordinates for vertex {:?}: {e}",
-                            vh.vertex_key()
-                        ),
-                    }
-                })?;
-                if coords.len() < 2 {
-                    return Err(CdtError::ValidationFailed {
-                        check: "foliation_assignment".to_string(),
-                        detail: format!(
-                            "vertex {:?} has {} coordinates, expected ≥ 2",
-                            vh.vertex_key(),
-                            coords.len()
-                        ),
-                    });
-                }
-                Ok((vh, coords[1]))
-            })
-            .collect::<CdtResult<Vec<_>>>()?;
-
-        let y_min = y_coords
-            .iter()
-            .map(|(_, y)| *y)
-            .fold(f64::INFINITY, f64::min);
-        let y_max = y_coords
-            .iter()
-            .map(|(_, y)| *y)
-            .fold(f64::NEG_INFINITY, f64::max);
-
-        let range = y_max - y_min;
-        let band_height = if range.abs() < f64::EPSILON {
-            1.0
-        } else {
-            range / f64::from(num_slices)
-        };
-
-        // Plan per-vertex labels in memory first so foliation validation errors
-        // do not partially mutate backend state.
-        let mut assignments = Vec::with_capacity(y_coords.len());
-        let mut slice_sizes = vec![0usize; num_slices as usize];
-        for (vh, y) in &y_coords {
-            let t = if range.abs() < f64::EPSILON {
-                0
-            } else {
-                let band_index = ((y - y_min) / band_height).floor();
-                f64_band_to_u32(band_index, num_slices - 1)
-            };
-            assignments.push((vh.vertex_key(), t));
-            slice_sizes[t as usize] += 1;
-        }
-
-        Self::check_time_slices(self.metadata.topology, num_slices)?;
-
-        let foliation =
-            Foliation::from_slice_sizes(slice_sizes, num_slices).map_err(CdtError::from)?;
-
-        // Clear stale cell classifications from any previous classify_all_cells() call,
-        // since vertex time labels are about to change.
-        let face_keys: Vec<_> = self.geometry.faces().map(|f| f.cell_key()).collect();
-        let previous_cell_data: Vec<_> = face_keys
-            .iter()
-            .map(|&key| (key, self.geometry.cell_data_by_key(key)))
-            .collect();
-        let previous_vertex_data: Vec<_> = assignments
-            .iter()
-            .map(|&(key, _)| (key, self.geometry.vertex_data_by_key(key)))
-            .collect();
-
-        let rollback_payloads = |geometry: &mut DelaunayBackend2D| -> Vec<String> {
-            let mut rollback_errors = Vec::new();
-
-            for &(key, data) in &previous_cell_data {
-                if let Err(err) = geometry.set_cell_data_by_key(key, data) {
-                    rollback_errors.push(format!("face {key:?}: {err}"));
-                }
-            }
-
-            for &(key, data) in &previous_vertex_data {
-                if let Err(err) = geometry.set_vertex_data_by_key(key, data) {
-                    rollback_errors.push(format!("vertex {key:?}: {err}"));
-                }
-            }
-
-            rollback_errors
-        };
-
-        for &key in &face_keys {
-            if let Err(err) = self.geometry.set_cell_data_by_key(key, None) {
-                let operation = "set_cell_data_by_key".to_string();
-                let target = format!("face {key:?}");
-                let detail = err.to_string();
-                let rollback_errors = rollback_payloads(&mut self.geometry);
-                return if rollback_errors.is_empty() {
-                    Err(CdtError::BackendMutationFailed {
-                        operation,
-                        target,
-                        detail,
-                    })
-                } else {
-                    Err(CdtError::BackendRollbackFailed {
-                        operation,
-                        target,
-                        detail,
-                        rollback_errors: rollback_errors.join("; "),
-                    })
-                };
-            }
-        }
-
-        // Write time labels directly to vertex data via set_vertex_data_by_key (O(1) per vertex).
-        for (vertex_key, t) in assignments {
-            if let Err(err) = self.geometry.set_vertex_data_by_key(vertex_key, Some(t)) {
-                let operation = "set_vertex_data_by_key".to_string();
-                let target = format!("vertex {vertex_key:?}");
-                let detail = format!("failed while assigning time label {t}: {err}");
-                let rollback_errors = rollback_payloads(&mut self.geometry);
-                return if rollback_errors.is_empty() {
-                    Err(CdtError::BackendMutationFailed {
-                        operation,
-                        target,
-                        detail,
-                    })
-                } else {
-                    Err(CdtError::BackendRollbackFailed {
-                        operation,
-                        target,
-                        detail,
-                        rollback_errors: rollback_errors.join("; "),
-                    })
-                };
-            }
-        }
-
-        self.metadata.time_slices = num_slices;
-        self.bump_modification_count();
-        self.foliation = Some(foliation);
-        self.mark_foliation_synchronized();
-        Ok(())
-    }
-
-    // -------------------------------------------------------------------------
-    // Foliation queries
-    // -------------------------------------------------------------------------
-
-    /// Returns `true` if this triangulation has an assigned foliation.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::CdtTriangulation;
-    ///
-    /// let tri = CdtTriangulation::from_seeded_points(5, 2, 2, 53).unwrap();
-    /// assert!(!tri.has_foliation());
-    /// ```
-    #[must_use]
-    pub fn has_foliation(&self) -> bool {
-        self.has_current_foliation()
-    }
-
-    /// Returns a reference to the foliation, if present.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::CdtTriangulation;
-    ///
-    /// let mut tri = CdtTriangulation::from_seeded_points(8, 2, 2, 53).unwrap();
-    /// tri.assign_foliation_by_y(2).unwrap();
-    /// assert!(tri.foliation().is_some());
-    /// ```
-    #[must_use]
-    pub fn foliation(&self) -> Option<&Foliation> {
-        if self.has_current_foliation() {
-            self.foliation.as_ref()
-        } else {
-            None
-        }
-    }
-
-    /// Returns the time slice label for a vertex, or `None` if no foliation
-    /// is present or the vertex is unlabeled.
-    ///
-    /// Reads the time label directly from the vertex data stored in the
-    /// Delaunay triangulation (like CDT++ `vertex->info()`).
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::prelude::geometry::*;
-    /// use causal_triangulations::CdtTriangulation;
-    ///
-    /// let mut tri = CdtTriangulation::from_seeded_points(8, 2, 2, 53).unwrap();
-    /// tri.assign_foliation_by_y(2).unwrap();
-    /// let vertex = tri.geometry().vertices().next().unwrap();
-    /// assert!(tri.time_label(&vertex).is_some());
-    /// ```
-    #[must_use]
-    pub fn time_label(&self, vertex: &DelaunayVertexHandle) -> Option<u32> {
-        self.foliation.as_ref()?;
-        self.geometry.vertex_data_by_key(vertex.vertex_key())
-    }
-
-    /// Returns all vertex handles that belong to time slice `t`.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::CdtTriangulation;
-    ///
-    /// let mut tri = CdtTriangulation::from_seeded_points(8, 2, 2, 53).unwrap();
-    /// tri.assign_foliation_by_y(2).unwrap();
-    /// assert!(!tri.vertices_at_time(0).is_empty());
-    /// ```
-    #[must_use]
-    pub fn vertices_at_time(&self, t: u32) -> Vec<DelaunayVertexHandle> {
-        if self.foliation.is_none() {
-            return vec![];
-        }
-        self.geometry
-            .vertices()
-            .filter(|vh| self.geometry.vertex_data_by_key(vh.vertex_key()) == Some(t))
-            .collect()
-    }
-
-    /// Returns per-slice vertex counts, or an empty slice if no foliation.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::CdtTriangulation;
-    ///
-    /// let mut tri = CdtTriangulation::from_seeded_points(8, 2, 2, 53).unwrap();
-    /// tri.assign_foliation_by_y(2).unwrap();
-    /// assert_eq!(tri.slice_sizes().len(), 2);
-    /// ```
-    #[must_use]
-    pub fn slice_sizes(&self) -> &[usize] {
-        self.foliation().map_or(&[], Foliation::slice_sizes)
-    }
-
-    /// Counts strict CDT triangles by the time slab of their lower slice.
-    ///
-    /// The returned vector has one entry per configured time slice.  Entry
-    /// `t` counts finite triangular cells whose [`CellType`] spans slice `t`
-    /// to `t + 1`; for toroidal triangulations the final entry includes cells
-    /// spanning the temporal wrap from `T - 1` to `0`.  If no current
-    /// foliation is present, returns an empty profile.
-    ///
-    /// Faces that cannot be classified as strict CDT cells are skipped, so the
-    /// profile sum can be smaller than [`Self::face_count`] for approximate or
-    /// partially foliated triangulations.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::prelude::triangulation::*;
-    ///
-    /// let tri = CdtTriangulation::from_cdt_strip(4, 3)
-    ///     .expect("build explicit CDT strip");
-    /// assert_eq!(tri.volume_profile(), vec![6, 6, 0]);
-    /// ```
-    #[must_use]
-    pub fn volume_profile(&self) -> Vec<u32> {
-        if !self.has_current_foliation() {
-            return Vec::new();
-        }
-
-        let Ok(slice_count) = usize::try_from(self.metadata.time_slices) else {
-            return Vec::new();
-        };
-        let mut profile = vec![0_u32; slice_count];
-
-        for face in self.geometry.faces() {
-            let Some(slice) = self.face_time_slice(&face) else {
-                continue;
-            };
-            let Ok(index) = usize::try_from(slice) else {
-                continue;
-            };
-            if let Some(count) = profile.get_mut(index) {
-                *count = count.saturating_add(1);
-            }
-        }
-
-        profile
-    }
-
-    // -------------------------------------------------------------------------
-    // Cell (triangle) classification
-    // -------------------------------------------------------------------------
-
-    /// Computes the temporal step distance between two time labels.
-    ///
-    /// For `OpenBoundary` topology this is just `t0.abs_diff(t1)`; for
-    /// `Toroidal` topology it is the circular distance
-    /// `min(d, T − d)` where `T = time_slices`, so the wrap-around edge
-    /// between slice `T − 1` and slice `0` reads as distance `1`.
-    ///
-    /// Out-of-range labels (`t ≥ time_slices`) bypass the toroidal wrap and
-    /// fall through to the raw difference so callers (e.g. validators) can
-    /// still detect them as causality violations rather than silently folding
-    /// `t == time_slices` into distance 0 via `min(T, 0)`.
-    #[must_use]
-    fn time_step_distance(&self, t0: u32, t1: u32) -> u32 {
-        let raw = t0.abs_diff(t1);
-        if matches!(self.metadata.topology, CdtTopology::Toroidal) {
-            let total = self.metadata.time_slices;
-            if total > 0 && t0 < total && t1 < total {
-                return raw.min(total - raw);
-            }
-        }
-        raw
-    }
-
-    /// Topology-aware variant of [`crate::cdt::foliation::classify_cell`].
-    ///
-    /// Uses [`Self::time_step_distance`] so the spacelike/timelike split
-    /// honours toroidal wrap.  Distinguishes Up vs Down by checking whether
-    /// the apex sits at `base + 1` or `base - 1` modulo `time_slices`.
-    fn classify_cell_with_topology(&self, t0: u32, t1: u32, t2: u32) -> Option<CellType> {
-        let mut dists = [
-            self.time_step_distance(t0, t1),
-            self.time_step_distance(t1, t2),
-            self.time_step_distance(t0, t2),
-        ];
-        dists.sort_unstable();
-        if dists != [0, 1, 1] {
-            return None;
-        }
-
-        let (base_slice, apex_slice) = if t0 == t1 {
-            (t0, t2)
-        } else if t1 == t2 {
-            (t1, t0)
-        } else if t0 == t2 {
-            (t0, t1)
-        } else {
-            return None;
-        };
-
-        let total = self.metadata.time_slices;
-        let toroidal = matches!(self.metadata.topology, CdtTopology::Toroidal) && total > 0;
-        let up_apex = if toroidal {
-            (base_slice + 1) % total
-        } else {
-            base_slice.checked_add(1)?
-        };
-        let down_apex = if toroidal {
-            if base_slice == 0 {
-                total - 1
-            } else {
-                base_slice - 1
-            }
-        } else {
-            base_slice.checked_sub(1)?
-        };
-        if apex_slice == up_apex {
-            Some(CellType::Up)
-        } else if apex_slice == down_apex {
-            Some(CellType::Down)
-        } else {
-            None
-        }
-    }
-
-    /// Returns the lower time-slab index assigned to a classifiable CDT face.
-    fn face_time_slice(&self, face: &DelaunayFaceHandle) -> Option<u32> {
-        self.cell_type(face)?;
-
-        let vertices = self.geometry.face_vertices(face).ok()?;
-        let [v0, v1, v2] = vertices.as_slice() else {
-            return None;
-        };
-
-        let labels = [
-            self.geometry.vertex_data_by_key(v0.vertex_key())?,
-            self.geometry.vertex_data_by_key(v1.vertex_key())?,
-            self.geometry.vertex_data_by_key(v2.vertex_key())?,
-        ];
-
-        match self.metadata.topology {
-            CdtTopology::OpenBoundary => Some(labels[0].min(labels[1]).min(labels[2])),
-            CdtTopology::Toroidal => {
-                let total = self.metadata.time_slices;
-                for slice in 0..total {
-                    let next = (slice + 1) % total;
-                    let spans_slab = labels.iter().all(|&label| label == slice || label == next)
-                        && labels.contains(&slice)
-                        && labels.contains(&next);
-                    if spans_slab {
-                        return Some(slice);
-                    }
-                }
-                None
-            }
-        }
-    }
-
-    /// Returns the causal classification of an edge from endpoint time labels.
-    ///
-    /// Returns `None` if no foliation is present, the edge endpoints cannot be
-    /// resolved, or either endpoint is missing a time label.  Distance is
-    /// circular when the topology is `Toroidal`.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::prelude::geometry::*;
-    /// use causal_triangulations::prelude::triangulation::*;
-    ///
-    /// let dt = build_delaunay2_with_data(&[
-    ///     ([0.0, 0.0], 0),
-    ///     ([1.0, 0.0], 0),
-    ///     ([0.5, 1.0], 1),
-    /// ])
-    /// .expect("build labeled triangle");
-    /// let backend = DelaunayBackend2D::from_triangulation(dt);
-    /// let tri = CdtTriangulation::from_labeled_delaunay(backend, 2, 2)
-    ///     .expect("create foliated triangulation");
-    ///
-    /// let edge = tri.geometry().edges().next().expect("triangle has edges");
-    /// let edge_type = tri.edge_type(&edge).expect("edge should be classifiable");
-    /// assert!(!matches!(edge_type, EdgeType::Acausal));
-    /// ```
-    #[must_use]
-    pub fn edge_type(&self, edge: &DelaunayEdgeHandle) -> Option<EdgeType> {
-        self.foliation.as_ref()?;
-
-        let (v0, v1) = self.geometry.edge_endpoints(edge)?;
-        let t0 = self.geometry.vertex_data_by_key(v0.vertex_key())?;
-        let t1 = self.geometry.vertex_data_by_key(v1.vertex_key())?;
-
-        Some(match self.time_step_distance(t0, t1) {
-            0 => EdgeType::Spacelike,
-            1 => EdgeType::Timelike,
-            _ => EdgeType::Acausal,
-        })
-    }
-
-    /// Classifies a triangle as Up (2,1) or Down (1,2) from vertex time labels.
-    ///
-    /// Returns `None` if no foliation is present, the face vertices cannot
-    /// be resolved, any vertex lacks a time label, or the triangle does not
-    /// span exactly one time slice (e.g. a boundary same-slice triangle).
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::prelude::triangulation::*;
-    ///
-    /// let mut tri = CdtTriangulation::from_seeded_points(12, 3, 2, 42)
-    ///     .expect("create seeded triangulation");
-    /// tri.assign_foliation_by_y(3)
-    ///     .expect("assign foliation from y-coordinates");
-    /// // Most cells should be classifiable; boundary cells may not be.
-    /// let classified: usize = tri.geometry().faces()
-    ///     .filter(|f| tri.cell_type(f).is_some())
-    ///     .count();
-    /// assert!(classified > 0);
-    /// ```
-    #[must_use]
-    pub fn cell_type(&self, face: &DelaunayFaceHandle) -> Option<CellType> {
-        self.foliation.as_ref()?;
-        let verts = self.geometry.face_vertices(face).ok()?;
-        if verts.len() != 3 {
-            return None;
-        }
-        let t0 = self.geometry.vertex_data_by_key(verts[0].vertex_key())?;
-        let t1 = self.geometry.vertex_data_by_key(verts[1].vertex_key())?;
-        let t2 = self.geometry.vertex_data_by_key(verts[2].vertex_key())?;
-        match self.metadata.topology {
-            CdtTopology::Toroidal => self.classify_cell_with_topology(t0, t1, t2),
-            CdtTopology::OpenBoundary => classify_cell(Some(t0), Some(t1), Some(t2)),
-        }
-    }
-
-    /// Reads the stored cell type from cell data, if previously classified.
-    ///
-    /// Returns `None` if the face has no cell data or the data does not
-    /// encode a valid [`CellType`].
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::prelude::geometry::*;
-    /// use causal_triangulations::CdtTriangulation;
-    ///
-    /// let mut tri = CdtTriangulation::from_cdt_strip(4, 2).unwrap();
-    /// tri.classify_all_cells().unwrap();
-    /// let face = tri.geometry().faces().next().unwrap();
-    /// let _stored = tri.cell_type_from_data(&face);
-    /// ```
-    #[must_use]
-    pub fn cell_type_from_data(&self, face: &DelaunayFaceHandle) -> Option<CellType> {
-        self.foliation()?;
-        let raw = self.geometry.cell_data_by_key(face.cell_key())?;
-        CellType::from_i32(raw)
-    }
-
-    /// Returns the edge classification for a triangular face.
-    ///
-    /// Edge ordering matches the face vertex cycle `(v0-v1, v1-v2, v2-v0)`.
-    /// Returns `None` if foliation is absent, face vertices cannot be resolved,
-    /// the face is not triangular, or any vertex is unlabeled.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::prelude::geometry::*;
-    /// use causal_triangulations::prelude::triangulation::*;
-    ///
-    /// let dt = build_delaunay2_with_data(&[
-    ///     ([0.0, 0.0], 0),
-    ///     ([1.0, 0.0], 0),
-    ///     ([0.5, 1.0], 1),
-    /// ])
-    /// .expect("build labeled triangle");
-    /// let backend = DelaunayBackend2D::from_triangulation(dt);
-    /// let tri = CdtTriangulation::from_labeled_delaunay(backend, 2, 2)
-    ///     .expect("create foliated triangulation");
-    ///
-    /// let face = tri.geometry().faces().next().expect("triangle has a face");
-    /// let edge_types = tri
-    ///     .face_edge_types(&face)
-    ///     .expect("face edge types should be available");
-    ///
-    /// let spacelike = edge_types
-    ///     .iter()
-    ///     .filter(|e| matches!(e, EdgeType::Spacelike))
-    ///     .count();
-    /// let timelike = edge_types
-    ///     .iter()
-    ///     .filter(|e| matches!(e, EdgeType::Timelike))
-    ///     .count();
-    /// assert_eq!(spacelike, 1);
-    /// assert_eq!(timelike, 2);
-    /// ```
-    #[must_use]
-    pub fn face_edge_types(&self, face: &DelaunayFaceHandle) -> Option<[EdgeType; 3]> {
-        self.foliation.as_ref()?;
-
-        let verts = self.geometry.face_vertices(face).ok()?;
-        if verts.len() != 3 {
-            return None;
-        }
-
-        let t = [
-            self.geometry.vertex_data_by_key(verts[0].vertex_key())?,
-            self.geometry.vertex_data_by_key(verts[1].vertex_key())?,
-            self.geometry.vertex_data_by_key(verts[2].vertex_key())?,
-        ];
-
-        let edge_classify = |a: u32, b: u32| -> Option<EdgeType> {
-            Some(match self.time_step_distance(a, b) {
-                0 => EdgeType::Spacelike,
-                1 => EdgeType::Timelike,
-                _ => EdgeType::Acausal,
-            })
-        };
-
-        Some([
-            edge_classify(t[0], t[1])?,
-            edge_classify(t[1], t[2])?,
-            edge_classify(t[2], t[0])?,
-        ])
-    }
-
-    /// Validates that every finite face has a strict CDT cell classification.
-    ///
-    /// If no foliation is present, succeeds vacuously. Otherwise every face
-    /// must classify as [`CellType::Up`] or [`CellType::Down`].
-    ///
-    /// # Errors
-    ///
-    /// Returns [`CdtError::ValidationFailed`] if any face is unclassifiable.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::prelude::triangulation::*;
-    ///
-    /// let tri = CdtTriangulation::from_cdt_strip(4, 2)
-    ///     .expect("build explicit CDT strip");
-    /// tri.validate_cell_classification()
-    ///     .expect("all strip cells classify");
-    /// ```
-    pub fn validate_cell_classification(&self) -> CdtResult<()> {
-        if self.foliation.is_none() {
-            return Ok(());
-        }
-
-        for face in self.geometry.faces() {
-            if self.cell_type(&face).is_none() {
-                return Err(CdtError::ValidationFailed {
-                    check: "cell_classification".to_string(),
-                    detail: format!(
-                        "face {:?} is not a strict CDT cell (expected Up or Down)",
-                        face.cell_key()
-                    ),
-                });
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Classifies every triangle and stores the result as cell data.
-    ///
-    /// Each classifiable cell receives `Some(CellType::to_i32())` via
-    /// `set_cell_data`.  On a foliated triangulation, all finite faces must be
-    /// classifiable; unclassifiable same-slice or multi-slice triangles are
-    /// reported as validation failures.
-    ///
-    /// Requires a foliation to be present; returns `Ok(None)` if there is none.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`CdtError::ValidationFailed`] if any foliated face is not an
-    /// Up or Down CDT triangle. Returns [`CdtError::BackendMutationFailed`] if
-    /// writing cell payloads to the backend fails, or
-    /// [`CdtError::BackendRollbackFailed`] if restoring previous payloads also fails.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::prelude::triangulation::*;
-    ///
-    /// let mut tri = CdtTriangulation::from_cdt_strip(4, 2)
-    ///     .expect("create explicit strip");
-    /// let classified = tri
-    ///     .classify_all_cells()
-    ///     .expect("classify cells")
-    ///     .expect("foliation is present");
-    /// assert!(classified > 0);
-    /// ```
-    pub fn classify_all_cells(&mut self) -> CdtResult<Option<usize>> {
-        if self.foliation.is_none() {
-            return Ok(None);
-        }
-
-        let faces: Vec<_> = self.geometry.faces().collect();
-        let mut classifications = Vec::with_capacity(faces.len());
-        for face in &faces {
-            let Some(ct) = self.cell_type(face) else {
-                return Err(CdtError::ValidationFailed {
-                    check: "cell_classification".to_string(),
-                    detail: format!(
-                        "face {:?} is not a strict CDT cell (expected Up or Down)",
-                        face.cell_key()
-                    ),
-                });
-            };
-            classifications.push((face.cell_key(), ct));
-        }
-
-        let count = classifications.len();
-        let previous_cell_data: Vec<_> = faces
-            .iter()
-            .map(|face| {
-                let key = face.cell_key();
-                (key, self.geometry.cell_data_by_key(key))
-            })
-            .collect();
-        let rollback_cell_payloads = |geometry: &mut DelaunayBackend2D| -> Vec<String> {
-            let mut rollback_errors = Vec::new();
-
-            for &(key, data) in &previous_cell_data {
-                if let Err(err) = geometry.set_cell_data_by_key(key, data) {
-                    rollback_errors.push(format!("face {key:?}: {err}"));
-                }
-            }
-
-            rollback_errors
-        };
-
-        // Clear all cell data first, then write fresh classifications.
-        for face in &faces {
-            let key = face.cell_key();
-            if let Err(err) = self.geometry.set_cell_data_by_key(key, None) {
-                let operation = "set_cell_data_by_key".to_string();
-                let target = format!("face {key:?}");
-                let detail =
-                    format!("failed to clear existing cell payload before classification: {err}");
-                let rollback_errors = rollback_cell_payloads(&mut self.geometry);
-                return if rollback_errors.is_empty() {
-                    Err(CdtError::BackendMutationFailed {
-                        operation,
-                        target,
-                        detail,
-                    })
-                } else {
-                    Err(CdtError::BackendRollbackFailed {
-                        operation,
-                        target,
-                        detail,
-                        rollback_errors: rollback_errors.join("; "),
-                    })
-                };
-            }
-        }
-        for (key, ct) in classifications {
-            if let Err(err) = self.geometry.set_cell_data_by_key(key, Some(ct.to_i32())) {
-                let operation = "set_cell_data_by_key".to_string();
-                let target = format!("face {key:?}");
-                let detail = format!(
-                    "failed to store classified cell payload {}: {err}",
-                    ct.to_i32()
-                );
-                let rollback_errors = rollback_cell_payloads(&mut self.geometry);
-                return if rollback_errors.is_empty() {
-                    Err(CdtError::BackendMutationFailed {
-                        operation,
-                        target,
-                        detail,
-                    })
-                } else {
-                    Err(CdtError::BackendRollbackFailed {
-                        operation,
-                        target,
-                        detail,
-                        rollback_errors: rollback_errors.join("; "),
-                    })
-                };
-            }
-        }
-        Ok(Some(count))
-    }
-
-    /// Rebuilds foliation bookkeeping from live backend vertex labels after a topology edit.
-    ///
-    /// Ergodic moves preserve labels on existing vertices and assign labels to
-    /// inserted vertices before calling this helper. The backend topology has
-    /// already been mutated, so this method only refreshes CDT-side aggregate
-    /// bookkeeping and cell payload classifications.
-    pub(crate) fn synchronize_foliation_from_live_labels(&mut self) -> CdtResult<()> {
-        if self.foliation.is_none() {
-            return Ok(());
-        }
-
-        let slice_sizes =
-            Self::live_slice_sizes_from_vertex_labels(&self.geometry, self.metadata.time_slices)?;
-        let foliation = Foliation::from_slice_sizes(slice_sizes, self.metadata.time_slices)
-            .map_err(CdtError::from)?;
-
-        self.foliation = Some(foliation);
-        match self.classify_all_cells() {
-            Ok(_) => {
-                self.mark_foliation_synchronized();
-                Ok(())
-            }
-            Err(err) => {
-                self.foliation = None;
-                self.foliation_synced_at_modification = None;
-                Err(err)
-            }
-        }
-    }
-
-    /// Validate CDT properties (geometry, Delaunay, topology, causality, foliation).
-    ///
-    /// # Errors
-    /// Returns error if any validation check fails.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::CdtTriangulation;
-    ///
-    /// let tri = CdtTriangulation::from_seeded_points(5, 2, 2, 53).unwrap();
-    /// assert!(tri.validate().is_ok());
-    /// ```
-    pub fn validate(&self) -> CdtResult<()> {
-        if !self.geometry.is_valid() {
-            return Err(CdtError::ValidationFailed {
-                check: "geometry".to_string(),
-                detail: format!(
-                    "triangulation is not valid (V={}, E={}, F={})",
-                    self.geometry.vertex_count(),
-                    self.geometry.edge_count(),
-                    self.geometry.face_count(),
-                ),
-            });
-        }
-
-        if !self.geometry.is_delaunay() {
-            return Err(CdtError::ValidationFailed {
-                check: "Delaunay".to_string(),
-                detail: format!(
-                    "triangulation does not satisfy Delaunay property (V={}, E={}, F={})",
-                    self.geometry.vertex_count(),
-                    self.geometry.edge_count(),
-                    self.geometry.face_count(),
-                ),
-            });
-        }
-
-        self.validate_topology()?;
-        self.validate_foliation()?;
-        self.validate_causality()?;
-        self.validate_cell_classification()?;
-
-        Ok(())
-    }
-
-    /// Validate causality constraints.
-    ///
-    /// If no foliation is present, succeeds vacuously (no causal structure
-    /// to check).  Otherwise delegates to [`validate_causality_delaunay`](Self::validate_causality_delaunay).
-    ///
-    /// # Errors
-    ///
-    /// Returns error if any edge spans more than one time slice (`|Δt| > 1`).
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::prelude::geometry::*;
-    /// use causal_triangulations::prelude::triangulation::*;
-    ///
-    /// let dt = build_delaunay2_with_data(&[
-    ///     ([0.0, 0.0], 0),
-    ///     ([1.0, 0.0], 0),
-    ///     ([0.5, 1.0], 1),
-    /// ])
-    /// .expect("build labeled triangle");
-    /// let backend = DelaunayBackend2D::from_triangulation(dt);
-    /// let tri = CdtTriangulation::from_labeled_delaunay(backend, 2, 2)
-    ///     .expect("create foliated triangulation");
-    /// assert!(tri.validate_causality().is_ok());
-    /// ```
-    pub fn validate_causality(&self) -> CdtResult<()> {
-        self.validate_causality_delaunay()
-    }
-
-    /// Validates the causal structure of this foliated triangulation.
-    ///
-    /// Reads time labels directly from face vertex data and checks that every
-    /// triangle lies within a single slice pair. In a 2D triangulation, this
-    /// implies that each edge of each finite face connects vertices within the
-    /// same slice or adjacent slices, while avoiding backend-specific edge
-    /// endpoint resolution.
-    ///
-    /// # Errors
-    ///
-    /// Returns error if any triangle spans more than one time slice, if a face
-    /// cannot be resolved to three vertices, or if any face vertex is unlabeled.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use causal_triangulations::prelude::geometry::*;
-    /// use causal_triangulations::prelude::triangulation::*;
-    ///
-    /// let dt = build_delaunay2_with_data(&[
-    ///     ([0.0, 0.0], 0),
-    ///     ([1.0, 0.0], 0),
-    ///     ([0.5, 1.0], 1),
-    /// ])
-    /// .expect("build labeled triangle");
-    /// let backend = DelaunayBackend2D::from_triangulation(dt);
-    /// let tri = CdtTriangulation::from_labeled_delaunay(backend, 2, 2)
-    ///     .expect("create foliated triangulation");
-    /// assert!(tri.validate_causality_delaunay().is_ok());
-    /// ```
-    #[expect(
-        clippy::too_many_lines,
-        reason = "causality validation includes detailed diagnostics for multiple face-resolution and label error paths"
-    )]
-    pub fn validate_causality_delaunay(&self) -> CdtResult<()> {
-        if self.foliation.is_none() {
-            return Ok(());
-        }
-
-        for face in self.geometry.faces() {
-            let verts = self.geometry.face_vertices(&face).map_err(|err| {
-                log::debug!(
-                    "Causality validation failed to resolve vertices for face {:?}: {err}; vertex_count={}, edge_count={}, face_count={}",
-                    face,
-                    self.geometry.vertex_count(),
-                    self.geometry.edge_count(),
-                    self.geometry.face_count(),
-                );
-                CdtError::ValidationFailed {
-                    check: "causality".to_string(),
-                    detail: "failed to resolve face vertices".to_string(),
-                }
-            })?;
-
-            if verts.len() != 3 {
-                return Err(CdtError::ValidationFailed {
-                    check: "causality".to_string(),
-                    detail: format!(
-                        "face {:?} has {} vertices, expected 3",
-                        face.cell_key(),
-                        verts.len(),
-                    ),
-                });
-            }
-
-            let t0 = self
-                .geometry
-                .vertex_data_by_key(verts[0].vertex_key())
-                .ok_or_else(|| {
-                    log::debug!(
-                        "Causality validation found unlabeled vertex {:?} while checking face {:?}",
-                        verts[0].vertex_key(),
-                        face,
-                    );
-                    CdtError::ValidationFailed {
-                        check: "causality".to_string(),
-                        detail: format!(
-                            "vertex {:?} has no time label in a foliated triangulation",
-                            verts[0].vertex_key(),
-                        ),
-                    }
-                })?;
-            let t1 = self
-                .geometry
-                .vertex_data_by_key(verts[1].vertex_key())
-                .ok_or_else(|| {
-                    log::debug!(
-                        "Causality validation found unlabeled vertex {:?} while checking face {:?}",
-                        verts[1].vertex_key(),
-                        face,
-                    );
-                    CdtError::ValidationFailed {
-                        check: "causality".to_string(),
-                        detail: format!(
-                            "vertex {:?} has no time label in a foliated triangulation",
-                            verts[1].vertex_key(),
-                        ),
-                    }
-                })?;
-            let t2 = self
-                .geometry
-                .vertex_data_by_key(verts[2].vertex_key())
-                .ok_or_else(|| {
-                    log::debug!(
-                        "Causality validation found unlabeled vertex {:?} while checking face {:?}",
-                        verts[2].vertex_key(),
-                        face,
-                    );
-                    CdtError::ValidationFailed {
-                        check: "causality".to_string(),
-                        detail: format!(
-                            "vertex {:?} has no time label in a foliated triangulation",
-                            verts[2].vertex_key(),
-                        ),
-                    }
-                })?;
-
-            // CDT triangle invariant: exactly 1 spacelike edge, 2 timelike edges.
-            // On Toroidal topology the wrap-around edge between slice T−1 and 0
-            // is timelike, so we use the topology-aware step distance.
-            let mut spacelike = 0;
-            let mut timelike = 0;
-
-            for (a, b) in [(t0, t1), (t1, t2), (t2, t0)] {
-                let step_distance = self.time_step_distance(a, b);
-                match step_distance {
-                    0 => spacelike += 1,
-                    1 => timelike += 1,
-                    _ => {
-                        return Err(CdtError::CausalityViolation {
-                            time_0: a.min(b),
-                            time_1: a.max(b),
-                            step_distance,
-                        });
-                    }
-                }
-            }
-
-            if !(spacelike == 1 && timelike == 2) {
-                return Err(CdtError::ValidationFailed {
-                    check: "causality".to_string(),
-                    detail: format!(
-                        "invalid CDT triangle at face {:?}: spacelike={}, timelike={}",
-                        face.cell_key(),
-                        spacelike,
-                        timelike
-                    ),
-                });
-            }
-        }
-
-        Ok(())
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -2635,163 +609,6 @@ mod tests {
         dimension: u8,
     ) -> CdtTriangulation<DelaunayBackend2D> {
         CdtTriangulation::wrap_unchecked(backend, time_slices, dimension, CdtTopology::OpenBoundary)
-    }
-
-    #[test]
-    fn test_remap_toroidal_generation_error_updates_context() {
-        let remapped = remap_toroidal_generation_error(
-            CdtError::DelaunayGenerationFailed {
-                vertex_count: 3,
-                coordinate_range: (-1.0, 1.0),
-                attempt: 7,
-                underlying_error: "builder failed".to_string(),
-            },
-            12,
-        );
-
-        assert!(matches!(
-            remapped,
-            CdtError::DelaunayGenerationFailed {
-                vertex_count: 12,
-                coordinate_range: (0.0, 1.0),
-                attempt: 1,
-                ref underlying_error,
-            } if underlying_error == "builder failed"
-        ));
-    }
-
-    #[test]
-    fn test_remap_toroidal_generation_error_preserves_other_errors() {
-        let original = CdtError::InvalidGenerationParameters {
-            issue: "bad".to_string(),
-            provided_value: "x".to_string(),
-            expected_range: "y".to_string(),
-        };
-        assert_eq!(
-            remap_toroidal_generation_error(original.clone(), 12),
-            original
-        );
-    }
-
-    #[test]
-    fn test_from_random_points() {
-        let triangulation =
-            CdtTriangulation::from_random_points(10, 3, 2).expect("Failed to create triangulation");
-
-        assert_eq!(triangulation.dimension(), 2);
-        assert_eq!(triangulation.time_slices(), 3);
-        assert!(triangulation.vertex_count() > 0);
-        assert!(triangulation.edge_count() > 0);
-        assert!(triangulation.face_count() > 0);
-    }
-
-    #[test]
-    fn test_from_random_points_various_sizes() {
-        let test_cases = [
-            (3, 1, "minimal"),
-            (5, 2, "small"),
-            (10, 3, "medium"),
-            (20, 5, "large"),
-        ];
-
-        for (vertices, time_slices, description) in test_cases {
-            let triangulation = CdtTriangulation::from_random_points(vertices, time_slices, 2)
-                .unwrap_or_else(|e| panic!("Failed to create {description} triangulation: {e}"));
-
-            assert_eq!(
-                triangulation.dimension(),
-                2,
-                "Dimension should be 2 for {description}"
-            );
-            assert_eq!(
-                triangulation.time_slices(),
-                time_slices,
-                "Time slices should match for {description}"
-            );
-            assert!(
-                triangulation.vertex_count() >= 3,
-                "Should have at least 3 vertices for {description}"
-            );
-            assert!(
-                triangulation.edge_count() > 0,
-                "Should have edges for {description}"
-            );
-            assert!(
-                triangulation.face_count() > 0,
-                "Should have faces for {description}"
-            );
-        }
-    }
-
-    #[test]
-    fn test_from_seeded_points() {
-        let seed = 42;
-        let triangulation = CdtTriangulation::from_seeded_points(8, 2, 2, seed)
-            .expect("Failed to create seeded triangulation");
-
-        assert_eq!(triangulation.dimension(), 2);
-        assert_eq!(triangulation.time_slices(), 2);
-        assert!(triangulation.vertex_count() > 0);
-        assert!(triangulation.edge_count() > 0);
-        assert!(triangulation.face_count() > 0);
-    }
-
-    #[test]
-    fn test_seeded_determinism() {
-        let seed = 123;
-        let params = (6, 3, 2);
-
-        let triangulation1 =
-            CdtTriangulation::from_seeded_points(params.0, params.1, params.2, seed)
-                .expect("Failed to create first triangulation");
-        let triangulation2 =
-            CdtTriangulation::from_seeded_points(params.0, params.1, params.2, seed)
-                .expect("Failed to create second triangulation");
-
-        // Should produce identical results
-        assert_eq!(triangulation1.vertex_count(), triangulation2.vertex_count());
-        assert_eq!(triangulation1.edge_count(), triangulation2.edge_count());
-        assert_eq!(triangulation1.face_count(), triangulation2.face_count());
-        assert_eq!(triangulation1.dimension(), triangulation2.dimension());
-        assert_eq!(triangulation1.time_slices(), triangulation2.time_slices());
-    }
-
-    #[test]
-    fn test_seeded_different_seeds() {
-        let params = (7, 2, 2);
-        let tri1 = CdtTriangulation::from_seeded_points(params.0, params.1, params.2, 456)
-            .expect("Failed to create triangulation with seed 456");
-        let tri2 = CdtTriangulation::from_seeded_points(params.0, params.1, params.2, 789)
-            .expect("Failed to create triangulation with seed 789");
-
-        // Both should succeed but may differ in structure
-        assert_eq!(tri1.dimension(), tri2.dimension());
-        assert_eq!(tri1.time_slices(), tri2.time_slices());
-        // Vertex counts should be same as input
-        assert_eq!(tri1.vertex_count(), 7);
-        assert_eq!(tri2.vertex_count(), 7);
-    }
-
-    #[test]
-    fn test_invalid_dimension() {
-        let invalid_dimensions = [0, 1, 3, 4, 5];
-        for dim in invalid_dimensions {
-            let result = CdtTriangulation::from_random_points(10, 3, dim);
-            assert!(result.is_err(), "Should fail with dimension {dim}");
-
-            if let Err(CdtError::UnsupportedDimension(d)) = result {
-                assert_eq!(d, u32::from(dim), "Error should report correct dimension");
-            } else {
-                panic!("Expected UnsupportedDimension error for dimension {dim}");
-            }
-        }
-    }
-
-    #[test]
-    fn test_from_seeded_points_rejects_invalid_dimension() {
-        let result = CdtTriangulation::from_seeded_points(10, 3, 3, 42);
-
-        assert!(matches!(result, Err(CdtError::UnsupportedDimension(3))));
     }
 
     #[test]
@@ -2849,62 +666,6 @@ mod tests {
                 && provided_value == "3"
                 && expected == "backend dimension (2)"
         ));
-    }
-
-    #[test]
-    fn test_from_seeded_points_rejects_zero_time_slices() {
-        let result = CdtTriangulation::from_seeded_points(5, 0, 2, 53);
-
-        assert!(matches!(
-            result,
-            Err(CdtError::InvalidTriangulationMetadata {
-                ref field,
-                ref provided_value,
-                ref expected,
-                ..
-            }) if field == "timeslices" && provided_value == "0" && expected == "≥ 1"
-        ));
-    }
-
-    #[test]
-    fn test_invalid_vertex_count() {
-        let invalid_counts = [0, 1, 2];
-        for count in invalid_counts {
-            let result = CdtTriangulation::from_random_points(count, 2, 2);
-            assert!(result.is_err(), "Should fail with {count} vertices");
-
-            match result {
-                Err(CdtError::InvalidGenerationParameters {
-                    issue,
-                    provided_value,
-                    ..
-                }) => {
-                    assert_eq!(issue, "Insufficient vertex count");
-                    assert_eq!(provided_value, count.to_string());
-                }
-                other => panic!(
-                    "Expected InvalidGenerationParameters for {count} vertices, got {other:?}"
-                ),
-            }
-        }
-    }
-
-    #[test]
-    fn test_invalid_vertex_count_seeded() {
-        let result = CdtTriangulation::from_seeded_points(2, 2, 2, 123);
-        assert!(result.is_err(), "Should fail with 2 vertices");
-
-        match result {
-            Err(CdtError::InvalidGenerationParameters {
-                issue,
-                provided_value,
-                ..
-            }) => {
-                assert_eq!(issue, "Insufficient vertex count");
-                assert_eq!(provided_value, "2");
-            }
-            other => panic!("Expected InvalidGenerationParameters, got {other:?}"),
-        }
     }
 
     #[test]
@@ -3061,21 +822,6 @@ mod tests {
     }
 
     #[test]
-    fn test_validation_success() {
-        // Use a known good seed that produces valid triangulation
-        const GOOD_SEED: u64 = 53; // Known to produce Euler=2
-
-        let triangulation = CdtTriangulation::from_seeded_points(5, 2, 2, GOOD_SEED)
-            .expect("Failed to create triangulation");
-
-        let result = triangulation.validate();
-        assert!(
-            result.is_ok(),
-            "Validation should succeed for good triangulation: {result:?}"
-        );
-    }
-
-    #[test]
     fn test_validate_topology() {
         // Test with various configurations to check topology validation
         let seeds = [53, 87, 203]; // Known good seeds that produce Euler=2
@@ -3090,26 +836,6 @@ mod tests {
                 "Topology validation should succeed for seed {seed}: {result:?}"
             );
         }
-    }
-
-    #[test]
-    fn test_validate_causality_no_foliation() {
-        let triangulation =
-            CdtTriangulation::from_random_points(5, 2, 2).expect("Failed to create triangulation");
-
-        // Without foliation, causality validation succeeds vacuously
-        let result = triangulation.validate_causality();
-        assert!(result.is_ok(), "Causality should pass without foliation");
-    }
-
-    #[test]
-    fn test_validate_foliation_no_foliation() {
-        let triangulation =
-            CdtTriangulation::from_random_points(5, 3, 2).expect("Failed to create triangulation");
-
-        // Without foliation, foliation validation succeeds vacuously
-        let result = triangulation.validate_foliation();
-        assert!(result.is_ok(), "Foliation should pass without foliation");
     }
 
     #[test]
@@ -3444,1196 +1170,6 @@ mod tests {
         );
     }
 
-    // =========================================================================
-    // Foliation tests
-    // =========================================================================
-
-    /// Builds an explicit strip and verifies it is a strict CDT mesh.
-    fn strict_strip(
-        vertices_per_slice: u32,
-        num_slices: u32,
-    ) -> CdtTriangulation<DelaunayBackend2D> {
-        let tri = CdtTriangulation::from_cdt_strip(vertices_per_slice, num_slices)
-            .expect("explicit strip construction should succeed");
-        assert_eq!(
-            tri.vertex_count(),
-            vertices_per_slice as usize * num_slices as usize
-        );
-        assert_eq!(
-            tri.face_count(),
-            2 * (vertices_per_slice as usize - 1) * (num_slices as usize - 1)
-        );
-        assert_eq!(
-            tri.slice_sizes(),
-            vec![vertices_per_slice as usize; num_slices as usize].as_slice()
-        );
-        tri.validate_foliation()
-            .expect("explicit strip foliation should validate");
-        tri.validate_causality_delaunay()
-            .expect("explicit strip causality should validate");
-        tri.validate_topology()
-            .expect("explicit strip topology should validate");
-        tri.validate_cell_classification()
-            .expect("all explicit strip cells should classify");
-        for face in tri.geometry().faces() {
-            assert!(tri.cell_type(&face).is_some());
-            assert!(tri.cell_type_from_data(&face).is_some());
-        }
-        tri
-    }
-
-    #[test]
-    fn test_from_cdt_strip_all_vertices_labeled() {
-        let tri = strict_strip(5, 3);
-        for vertex in tri.geometry().vertices() {
-            assert!(tri.time_label(&vertex).is_some());
-        }
-    }
-
-    #[test]
-    fn test_from_labeled_delaunay_preserves_foliation() {
-        let backend = labeled_triangle_backend([0, 0, 1]);
-
-        let tri = CdtTriangulation::from_labeled_delaunay(backend, 2, 2)
-            .expect("Should preserve labels as foliation");
-
-        assert!(tri.has_foliation());
-        assert_eq!(tri.slice_sizes(), &[2, 1]);
-        assert!(tri.validate_foliation().is_ok());
-
-        for vh in tri.geometry().vertices() {
-            assert!(tri.time_label(&vh).is_some());
-        }
-    }
-
-    #[test]
-    fn test_from_labeled_delaunay_rejects_invalid_dimension() {
-        let backend = labeled_triangle_backend([0, 0, 1]);
-
-        let result = CdtTriangulation::from_labeled_delaunay(backend, 2, 3);
-
-        assert!(matches!(result, Err(CdtError::UnsupportedDimension(3))));
-    }
-
-    #[test]
-    fn test_from_labeled_delaunay_rejects_zero_slices() {
-        let backend = labeled_triangle_backend([0, 0, 1]);
-
-        let result = CdtTriangulation::from_labeled_delaunay(backend, 0, 2);
-
-        assert!(matches!(
-            result,
-            Err(CdtError::InvalidTriangulationMetadata {
-                ref field,
-                ref provided_value,
-                ref expected,
-                ..
-            }) if field == "timeslices" && provided_value == "0" && expected == "≥ 1"
-        ));
-    }
-
-    #[test]
-    fn test_validate_foliation_missing_label() {
-        let backend = labeled_triangle_backend([0, 0, 1]);
-
-        let mut tri = CdtTriangulation::from_labeled_delaunay(backend, 2, 2)
-            .expect("Should preserve labels as foliation");
-
-        let vertex_to_clear = tri
-            .geometry()
-            .vertices()
-            .next()
-            .expect("Triangle should contain a vertex");
-
-        tri.set_vertex_data(&vertex_to_clear, None)
-            .expect("Expected valid vertex handle while clearing label");
-
-        assert!(
-            !tri.has_foliation(),
-            "CDT label mutation should invalidate cached foliation bookkeeping"
-        );
-        assert!(
-            tri.slice_sizes().is_empty(),
-            "stale foliation bookkeeping should not be exposed via slice_sizes()"
-        );
-
-        let result = tri.validate_foliation();
-        assert!(matches!(
-            result,
-            Err(CdtError::Foliation(FoliationError::MissingVertexLabel {
-                vertex: 0
-            }))
-        ));
-    }
-
-    #[test]
-    fn test_validate_foliation_out_of_range_label() {
-        let dt = build_delaunay2_with_data(&[([0.0, 0.0], 0), ([1.0, 0.0], 0), ([0.5, 1.0], 1)])
-            .expect("Should build labeled triangle");
-        let backend = DelaunayBackend2D::from_triangulation(dt);
-
-        let mut tri = CdtTriangulation::from_labeled_delaunay(backend, 2, 2)
-            .expect("Should preserve labels as foliation");
-
-        let vertex_to_mutate = tri
-            .geometry()
-            .vertices()
-            .next()
-            .expect("Triangle should contain a vertex");
-
-        tri.set_vertex_data(&vertex_to_mutate, Some(7))
-            .expect("Expected valid vertex handle while mutating label");
-
-        let result = tri.validate_foliation();
-        assert!(matches!(
-            result,
-            Err(CdtError::Foliation(FoliationError::OutOfRangeVertexLabel {
-                vertex: 0,
-                label: 7,
-                expected_range_end: 2,
-            }))
-        ));
-    }
-
-    #[test]
-    fn test_validate_foliation_slice_mismatch() {
-        let dt = build_delaunay2_with_data(&[([0.0, 0.0], 0), ([1.0, 0.0], 0), ([0.5, 1.0], 1)])
-            .expect("Should build labeled triangle");
-        let backend = DelaunayBackend2D::from_triangulation(dt);
-
-        let mut tri = CdtTriangulation::from_labeled_delaunay(backend, 2, 2)
-            .expect("Should preserve labels as foliation");
-
-        let vertex_to_move = tri
-            .geometry()
-            .vertices()
-            .find(|vh| tri.geometry().vertex_data_by_key(vh.vertex_key()) == Some(0))
-            .expect("Triangle should contain a vertex in slice 0");
-
-        tri.set_vertex_data(&vertex_to_move, Some(1))
-            .expect("Expected valid vertex handle while mutating label");
-
-        let result = tri.validate_foliation();
-        assert!(matches!(
-            result,
-            Err(CdtError::Foliation(FoliationError::LabelMismatch { .. }))
-        ));
-    }
-
-    #[test]
-    fn test_validate_foliation_rejects_stored_label_count_mismatch() {
-        let backend = labeled_triangle_backend([0, 0, 1]);
-        let mut tri = CdtTriangulation::from_labeled_delaunay(backend, 2, 2)
-            .expect("Should preserve labels as foliation");
-        tri.foliation = Some(
-            Foliation::from_slice_sizes(vec![1, 1], 2)
-                .expect("non-empty mismatched bookkeeping is constructible"),
-        );
-
-        let result = tri.validate_foliation();
-
-        assert!(matches!(
-            result,
-            Err(CdtError::Foliation(FoliationError::LabelCountMismatch {
-                labeled: 2,
-                expected: 3,
-            }))
-        ));
-    }
-
-    #[test]
-    fn test_from_labeled_delaunay_rejects_out_of_range_labels() {
-        let dt = build_delaunay2_with_data(&[([0.0, 0.0], 0), ([1.0, 0.0], 0), ([0.5, 1.0], 5)])
-            .expect("Should build labeled triangle");
-        let backend = DelaunayBackend2D::from_triangulation(dt);
-
-        let result = CdtTriangulation::from_labeled_delaunay(backend, 2, 2);
-        assert!(matches!(
-            result,
-            Err(CdtError::Foliation(FoliationError::OutOfRangeVertexLabel {
-                label: 5,
-                expected_range_end: 2,
-                ..
-            }))
-        ));
-    }
-
-    #[test]
-    fn test_from_labeled_delaunay_rejects_empty_intermediate_slice() {
-        let dt = build_delaunay2_with_data(&[([0.0, 0.0], 0), ([1.0, 0.0], 2), ([0.5, 1.0], 2)])
-            .expect("Should build labeled triangle");
-        let backend = DelaunayBackend2D::from_triangulation(dt);
-
-        let result = CdtTriangulation::from_labeled_delaunay(backend, 3, 2);
-        assert!(matches!(
-            result,
-            Err(CdtError::Foliation(FoliationError::EmptySlice { slice: 1 }))
-        ));
-    }
-
-    #[test]
-    fn test_from_cdt_strip_edge_classification() {
-        let tri = strict_strip(5, 3);
-        for edge in tri.geometry().edges() {
-            assert!(matches!(
-                tri.edge_type(&edge),
-                Some(EdgeType::Spacelike | EdgeType::Timelike)
-            ));
-        }
-    }
-
-    #[test]
-    fn test_from_cdt_strip_rejects_invalid_params() {
-        let few_vertices = CdtTriangulation::from_cdt_strip(3, 3);
-        assert!(matches!(
-            few_vertices,
-            Err(CdtError::InvalidGenerationParameters {
-                ref issue,
-                ref provided_value,
-                ref expected_range,
-            }) if issue == "Insufficient vertices per slice"
-                && provided_value == "3"
-                && expected_range == "≥ 4"
-        ));
-
-        let few_slices = CdtTriangulation::from_cdt_strip(4, 1);
-        assert!(matches!(
-            few_slices,
-            Err(CdtError::InvalidGenerationParameters {
-                ref issue,
-                ref provided_value,
-                ref expected_range,
-            }) if issue == "Insufficient number of time slices"
-                && provided_value == "1"
-                && expected_range == "≥ 2"
-        ));
-    }
-
-    #[test]
-    fn test_from_cdt_strip_rejects_cell_count_overflow() {
-        let result = CdtTriangulation::from_cdt_strip(65_535, 65_537);
-
-        assert!(matches!(
-            result,
-            Err(CdtError::InvalidGenerationParameters {
-                ref issue,
-                ref provided_value,
-                ref expected_range,
-            }) if issue == "Cell count overflow"
-                && provided_value == "2 × 4294836224"
-                && expected_range == "product ≤ u32::MAX"
-        ));
-    }
-
-    #[test]
-    fn test_from_cdt_strip_builds_valid_mesh() {
-        let tri = CdtTriangulation::from_cdt_strip(4, 2).expect("explicit strip should build");
-        assert_eq!(tri.vertex_count(), 8);
-        assert_eq!(tri.face_count(), 6);
-        assert!(tri.validate_topology().is_ok());
-        assert!(tri.validate_foliation().is_ok());
-        assert!(tri.validate_causality_delaunay().is_ok());
-        assert!(tri.validate_cell_classification().is_ok());
-    }
-
-    #[test]
-    fn test_explicit_strip_count_validation_rejects_face_mismatch() {
-        let tri = CdtTriangulation::from_cdt_strip(4, 2).expect("explicit strip should build");
-        let result = validate_strip_counts(tri.geometry(), 8, 7, 8, 7, 4, 2, 1.0);
-
-        assert!(matches!(
-            result,
-            Err(CdtError::DelaunayGenerationFailed {
-                vertex_count: 8,
-                coordinate_range: (0.0, 1.0),
-                attempt: 1,
-                ref underlying_error,
-            }) if underlying_error.contains("build_delaunay2_from_cells()/from_cdt_strip()")
-                && underlying_error.contains("produced 6 faces, expected 7")
-                && underlying_error.contains("vertices_per_slice=4")
-                && underlying_error.contains("num_slices=2")
-        ));
-    }
-
-    #[test]
-    fn test_vertices_at_time() {
-        let tri = strict_strip(4, 3);
-        for t in 0..3 {
-            assert_eq!(tri.vertices_at_time(t).len(), 4);
-        }
-    }
-
-    #[test]
-    fn test_assign_foliation_by_y() {
-        let mut tri = CdtTriangulation::from_seeded_points(15, 3, 2, 42)
-            .expect("Failed to create deterministic triangulation");
-
-        assert!(!tri.has_foliation());
-
-        tri.assign_foliation_by_y(3)
-            .expect("Should assign foliation");
-
-        assert!(tri.has_foliation());
-        assert_eq!(tri.time_slices(), 3);
-        assert_eq!(tri.slice_sizes().iter().sum::<usize>(), tri.vertex_count());
-
-        // All vertices should now be labeled
-        for vh in tri.geometry().vertices() {
-            assert!(tri.time_label(&vh).is_some());
-        }
-
-        // Foliation validation should pass
-        let result = tri.validate_foliation();
-        assert!(
-            result.is_ok(),
-            "Foliation validation should pass: {result:?}"
-        );
-    }
-
-    #[test]
-    fn test_assign_foliation_by_y_updates_metadata() {
-        let mut tri = CdtTriangulation::from_seeded_points(15, 3, 2, 42)
-            .expect("Failed to create deterministic triangulation");
-        let initial_last_modified = tri.metadata().last_modified;
-        let initial_modification_count = tri.metadata().modification_count;
-        let initial_edge_count = tri.edge_count();
-        let initial_euler_characteristic = tri.geometry().euler_characteristic();
-
-        thread::sleep(Duration::from_millis(5));
-
-        tri.assign_foliation_by_y(3)
-            .expect("Should update foliation metadata");
-
-        assert!(tri.metadata().last_modified > initial_last_modified);
-        assert_eq!(
-            tri.metadata().modification_count,
-            initial_modification_count + 1,
-            "Foliation assignment should count as a modification"
-        );
-        assert_eq!(tri.edge_count(), initial_edge_count);
-        assert_eq!(
-            tri.geometry().euler_characteristic(),
-            initial_euler_characteristic
-        );
-    }
-
-    #[test]
-    fn test_assign_foliation_by_y_invalidates_cache() {
-        let mut tri = CdtTriangulation::from_seeded_points(15, 3, 2, 42)
-            .expect("Failed to create deterministic triangulation");
-
-        tri.refresh_cache();
-        assert!(tri.cache.edge_count.is_some());
-        assert!(tri.cache.euler_char.is_some());
-
-        tri.assign_foliation_by_y(3)
-            .expect("Should invalidate cache when assigning foliation");
-
-        assert!(
-            tri.cache.edge_count.is_none(),
-            "assign_foliation_by_y should clear cached edge count via invalidate_cache()"
-        );
-        assert!(
-            tri.cache.euler_char.is_none(),
-            "assign_foliation_by_y should clear cached Euler characteristic via invalidate_cache()"
-        );
-    }
-
-    #[test]
-    fn test_assign_foliation_zero_slices() {
-        let mut tri = CdtTriangulation::from_random_points(5, 2, 2).unwrap();
-        assert!(tri.assign_foliation_by_y(0).is_err());
-    }
-
-    #[test]
-    fn test_assign_foliation_empty_slice_error_preserves_metadata() {
-        let mut tri =
-            CdtTriangulation::from_random_points(6, 2, 2).expect("Failed to create triangulation");
-        let initial_time_slices = tri.time_slices();
-        let initial_modification_count = tri.metadata().modification_count;
-        let vertex_keys: Vec<_> = tri
-            .geometry()
-            .vertices()
-            .map(|vh| vh.vertex_key())
-            .collect();
-
-        let requested_slices = u32::try_from(tri.vertex_count())
-            .expect("vertex count should fit into u32 for this test")
-            .saturating_add(1);
-        let result = tri.assign_foliation_by_y(requested_slices);
-
-        assert!(matches!(
-            result,
-            Err(CdtError::Foliation(FoliationError::EmptySlice { .. }))
-        ));
-        assert_eq!(tri.time_slices(), initial_time_slices);
-        assert_eq!(
-            tri.metadata().modification_count,
-            initial_modification_count
-        );
-        assert!(!tri.has_foliation());
-
-        for key in vertex_keys {
-            assert_eq!(
-                tri.geometry().vertex_data_by_key(key),
-                None,
-                "failed assignment should not write vertex labels"
-            );
-        }
-    }
-
-    #[test]
-    fn test_all_faces_are_valid_cdt_triangles() {
-        let tri = strict_strip(5, 3);
-        for face in tri.geometry().faces() {
-            let edge_types = tri
-                .face_edge_types(&face)
-                .expect("explicit strip face should expose edge types");
-            let spacelike = edge_types
-                .iter()
-                .filter(|edge_type| matches!(edge_type, EdgeType::Spacelike))
-                .count();
-            let timelike = edge_types
-                .iter()
-                .filter(|edge_type| matches!(edge_type, EdgeType::Timelike))
-                .count();
-            assert_eq!(spacelike, 1);
-            assert_eq!(timelike, 2);
-        }
-    }
-
-    #[test]
-    fn test_assign_foliation_single_slice() {
-        let mut tri =
-            CdtTriangulation::from_random_points(6, 1, 2).expect("Failed to create triangulation");
-
-        tri.assign_foliation_by_y(1)
-            .expect("Should assign single-slice foliation");
-
-        assert!(tri.has_foliation());
-        // All vertices should be in slice 0
-        for vh in tri.geometry().vertices() {
-            assert_eq!(tri.time_label(&vh), Some(0));
-        }
-        assert_eq!(tri.slice_sizes(), &[tri.vertex_count()]);
-    }
-
-    #[test]
-    fn test_no_foliation_queries_return_none() {
-        let tri = CdtTriangulation::from_random_points(5, 2, 2).unwrap();
-        assert!(!tri.has_foliation());
-        assert!(tri.foliation().is_none());
-        assert!(tri.slice_sizes().is_empty());
-        assert!(tri.vertices_at_time(0).is_empty());
-
-        // time_label and edge_type return None
-        for vh in tri.geometry().vertices() {
-            assert_eq!(tri.time_label(&vh), None);
-        }
-        for face in tri.geometry().faces() {
-            assert!(tri.face_edge_types(&face).is_none());
-        }
-    }
-
-    // =========================================================================
-    // Cell classification tests
-    // =========================================================================
-
-    #[test]
-    fn test_cell_type_returns_up_or_down() {
-        let tri = strict_strip(5, 3);
-        for face in tri.geometry().faces() {
-            assert!(matches!(
-                tri.cell_type(&face),
-                Some(CellType::Up | CellType::Down)
-            ));
-        }
-    }
-
-    #[test]
-    fn test_cell_type_no_foliation_returns_none() {
-        let tri = CdtTriangulation::from_random_points(5, 2, 2).unwrap();
-        assert!(!tri.has_foliation());
-
-        for face in tri.geometry().faces() {
-            assert_eq!(tri.cell_type(&face), None);
-        }
-    }
-
-    #[test]
-    fn test_classify_all_cells_stores_data() {
-        let mut tri = strict_strip(5, 3);
-        let classified = tri
-            .classify_all_cells()
-            .expect("strict strip cells should classify")
-            .expect("foliation is present");
-        assert_eq!(classified, tri.face_count());
-    }
-
-    #[test]
-    fn test_classify_all_cells_no_foliation_returns_none() {
-        let mut tri = CdtTriangulation::from_random_points(5, 2, 2).unwrap();
-        assert_eq!(
-            tri.classify_all_cells()
-                .expect("No foliation should classify as a no-op"),
-            None
-        );
-    }
-
-    #[test]
-    fn test_validate_cell_classification_no_foliation_succeeds() {
-        let tri = CdtTriangulation::from_random_points(5, 2, 2).unwrap();
-
-        tri.validate_cell_classification()
-            .expect("missing foliation should validate vacuously");
-    }
-
-    #[test]
-    fn test_validate_and_classify_use_stored_foliation_after_label_mutation() {
-        let mut tri = CdtTriangulation::from_cdt_strip(4, 2).expect("Should build CDT strip");
-        let vertex = tri
-            .geometry()
-            .vertices()
-            .next()
-            .expect("CDT strip should contain vertices");
-        let label = tri
-            .geometry()
-            .vertex_data_by_key(vertex.vertex_key())
-            .expect("CDT strip vertices should be labeled");
-
-        tri.set_vertex_data(&vertex, Some(label))
-            .expect("Expected valid vertex handle while preserving label");
-
-        assert!(
-            !tri.has_foliation(),
-            "CDT label mutation should invalidate synchronized foliation bookkeeping"
-        );
-        tri.validate_cell_classification()
-            .expect("stored foliation should still drive live cell validation");
-
-        let classified = tri
-            .classify_all_cells()
-            .expect("stored foliation should still drive live cell classification")
-            .expect("foliation is still present");
-        assert_eq!(classified, tri.face_count());
-    }
-
-    #[test]
-    fn test_cell_type_from_data_before_classify_returns_none() {
-        let dt = build_delaunay2_with_data(&[([0.0, 0.0], 0), ([1.0, 0.0], 0), ([0.5, 1.0], 1)])
-            .expect("Should build labeled triangle");
-        let backend = DelaunayBackend2D::from_triangulation(dt);
-        let tri = CdtTriangulation::from_labeled_delaunay(backend, 2, 2)
-            .expect("Should preserve labels as foliation");
-
-        let face = tri
-            .geometry()
-            .faces()
-            .next()
-            .expect("Triangle should contain a face");
-        assert_eq!(tri.cell_type_from_data(&face), None);
-    }
-
-    #[test]
-    fn test_cell_type_from_data_returns_none_when_foliation_stale() {
-        let dt = build_delaunay2_with_data(&[([0.0, 0.0], 0), ([1.0, 0.0], 0), ([0.5, 1.0], 1)])
-            .expect("Should build labeled triangle");
-        let backend = DelaunayBackend2D::from_triangulation(dt);
-        let mut tri = CdtTriangulation::from_labeled_delaunay(backend, 2, 2)
-            .expect("Should preserve labels as foliation");
-
-        let face = tri
-            .geometry()
-            .faces()
-            .next()
-            .expect("Triangle should contain a face");
-        let vertex_to_mutate = tri
-            .geometry()
-            .vertices()
-            .next()
-            .expect("Triangle should contain a vertex");
-
-        let classified = tri
-            .classify_all_cells()
-            .expect("Should classify cells with foliation")
-            .expect("Foliation is present");
-        assert!(classified > 0);
-        assert!(tri.cell_type_from_data(&face).is_some());
-
-        tri.set_vertex_data(&vertex_to_mutate, Some(7))
-            .expect("Expected valid vertex handle while mutating label");
-
-        assert_eq!(tri.cell_type_from_data(&face), None);
-    }
-
-    #[test]
-    fn test_classify_all_cells_on_labeled_triangle() {
-        let dt = build_delaunay2_with_data(&[([0.0, 0.0], 0), ([1.0, 0.0], 0), ([0.5, 1.0], 1)])
-            .expect("Should build labeled triangle");
-        let backend = DelaunayBackend2D::from_triangulation(dt);
-        let mut tri = CdtTriangulation::from_labeled_delaunay(backend, 2, 2)
-            .expect("Should preserve labels as foliation");
-
-        // cell_type should classify before bulk classification
-        let face = tri
-            .geometry()
-            .faces()
-            .next()
-            .expect("Triangle should contain a face");
-        let live_ct = tri
-            .cell_type(&face)
-            .expect("Single face should be classifiable");
-        assert!(
-            live_ct == CellType::Up || live_ct == CellType::Down,
-            "Single face spanning two slices must be Up or Down"
-        );
-
-        // Bulk classify and verify stored data matches live classification
-        let count = tri
-            .classify_all_cells()
-            .expect("classify_all_cells should succeed")
-            .expect("foliation is present");
-        assert_eq!(count, 1, "Single-face triangulation should classify 1 cell");
-
-        let stored_ct = tri
-            .cell_type_from_data(&face)
-            .expect("Stored cell type should be present after classification");
-        assert_eq!(
-            live_ct, stored_ct,
-            "Stored classification should match live classification"
-        );
-    }
-
-    #[test]
-    fn test_validate_cell_classification_rejects_same_slice_triangle() {
-        let dt = build_delaunay2_with_data(&[([0.0, 0.0], 0), ([1.0, 0.0], 0), ([0.5, 1.0], 0)])
-            .expect("Should build same-slice triangle");
-        let backend = DelaunayBackend2D::from_triangulation(dt);
-        let mut tri = CdtTriangulation::from_labeled_delaunay(backend, 1, 2)
-            .expect("single-slice labels should build foliation");
-
-        let validation = tri.validate_cell_classification();
-        assert!(matches!(
-            validation,
-            Err(CdtError::ValidationFailed { ref check, .. })
-                if check == "cell_classification"
-        ));
-        let classification = tri.classify_all_cells();
-        assert!(matches!(
-            classification,
-            Err(CdtError::ValidationFailed { ref check, .. })
-                if check == "cell_classification"
-        ));
-    }
-
-    #[test]
-    fn test_vertices_at_time_with_foliation() {
-        let mut tri = CdtTriangulation::from_seeded_points(15, 3, 2, 42)
-            .expect("Failed to create deterministic triangulation");
-
-        tri.assign_foliation_by_y(3)
-            .expect("Should assign foliation");
-
-        // Every slice should have at least one vertex
-        for t in 0..3 {
-            let verts = tri.vertices_at_time(t);
-            assert!(
-                !verts.is_empty(),
-                "Slice {t} should have at least one vertex"
-            );
-            // Each returned vertex should actually carry that time label
-            for vh in &verts {
-                assert_eq!(
-                    tri.time_label(vh),
-                    Some(t),
-                    "Vertex returned by vertices_at_time({t}) should have label {t}"
-                );
-            }
-        }
-
-        // Sum across all slices should equal total vertex count
-        let total: usize = (0..3).map(|t| tri.vertices_at_time(t).len()).sum();
-        assert_eq!(
-            total,
-            tri.vertex_count(),
-            "Sum of vertices_at_time across all slices should equal vertex_count"
-        );
-
-        // Non-existent slice should return empty
-        assert!(tri.vertices_at_time(999).is_empty());
-    }
-
-    #[test]
-    fn test_assign_foliation_by_y_reassignment() {
-        let mut tri =
-            CdtTriangulation::from_cdt_strip(5, 3).expect("Failed to create deterministic strip");
-
-        // First assignment: 3 slices
-        tri.assign_foliation_by_y(3)
-            .expect("First foliation assignment should succeed");
-        assert!(tri.has_foliation());
-        assert_eq!(tri.time_slices(), 3);
-        let sizes_3 = tri.slice_sizes().to_vec();
-        assert_eq!(sizes_3.len(), 3);
-
-        // Classify cells so we can verify stale data is cleared
-        let classified_before = tri
-            .classify_all_cells()
-            .expect("classify_all_cells should succeed")
-            .expect("foliation is present");
-        assert!(classified_before > 0);
-
-        // Re-assign with 2 slices
-        tri.assign_foliation_by_y(2)
-            .expect("Re-assignment with different slice count should succeed");
-        assert!(tri.has_foliation());
-        assert_eq!(tri.time_slices(), 2);
-        let sizes_2 = tri.slice_sizes().to_vec();
-        assert_eq!(sizes_2.len(), 2);
-        assert_eq!(
-            sizes_2.iter().sum::<usize>(),
-            tri.vertex_count(),
-            "Slice sizes should sum to vertex count after re-assignment"
-        );
-
-        // Stale cell data from the 3-slice classification should be cleared;
-        // cell_type_from_data should return None until classify_all_cells is
-        // called again with the new foliation.
-        for face in tri.geometry().faces() {
-            assert_eq!(
-                tri.cell_type_from_data(&face),
-                None,
-                "Stale cell data should be cleared after foliation re-assignment"
-            );
-        }
-
-        // Foliation and causality validation should still pass
-        assert!(
-            tri.validate_foliation().is_ok(),
-            "Foliation should be valid after re-assignment"
-        );
-    }
-
-    /// Builds stable diagnostic text for seeded-triangulation comparisons.
-    fn deterministic_triangle_debug_summary(backend: &DelaunayBackend2D) -> String {
-        let mut vertices: Vec<_> = backend
-            .vertices()
-            .map(|vh| {
-                let coords = backend.vertex_coordinates(&vh).map_or_else(
-                    |err| format!("coord_error:{err}"),
-                    |coords| format!("{coords:?}"),
-                );
-                format!(
-                    "{:?}@{}:{:?}",
-                    vh.vertex_key(),
-                    coords,
-                    backend.vertex_data_by_key(vh.vertex_key())
-                )
-            })
-            .collect();
-        vertices.sort_unstable();
-
-        let mut edges: Vec<_> = backend
-            .edges()
-            .map(|edge| match backend.edge_endpoints(&edge) {
-                Some((v0, v1)) => format!(
-                    "{:?}<->{:?}:{:?}->{:?}",
-                    v0.vertex_key(),
-                    v1.vertex_key(),
-                    backend.vertex_data_by_key(v0.vertex_key()),
-                    backend.vertex_data_by_key(v1.vertex_key())
-                ),
-                None => "endpoint_error:unavailable".to_string(),
-            })
-            .collect();
-        edges.sort_unstable();
-
-        format!(
-            "vertex_count={}, edge_count={}, face_count={}, is_valid={}, is_delaunay={}, vertices=[{}], edges=[{}]",
-            backend.vertex_count(),
-            backend.edge_count(),
-            backend.face_count(),
-            backend.is_valid(),
-            backend.is_delaunay(),
-            vertices.join(", "),
-            edges.join(", "),
-        )
-    }
-
-    // =========================================================================
-    // Causality violation detection
-    // =========================================================================
-
-    #[test]
-    fn test_causality_violation_detected() {
-        // Use a hand-built triangle so this test does not depend on Delaunay
-        // tie-breaking for a larger strip mesh.
-        let dt = build_delaunay2_with_data(&[([0.0, 0.0], 0), ([1.0, 0.0], 0), ([0.5, 1.0], 1)])
-            .expect("Should build deterministic causal triangle");
-        let backend = DelaunayBackend2D::from_triangulation(dt);
-        let mut tri = unchecked_open_boundary(backend, 2, 2);
-
-        // Derive the foliation from coordinates instead of relying on
-        // build_delaunay2_with_data() to preserve vertex data across platforms.
-        // This test targets causality validation, not builder label retention.
-        tri.assign_foliation_by_y(2)
-            .expect("Should derive foliation from triangle coordinates");
-
-        assert_eq!(
-            tri.slice_sizes(),
-            &[2, 1],
-            "Deterministic triangle should assign slice sizes [2, 1], got {:?}; {}",
-            tri.slice_sizes(),
-            deterministic_triangle_debug_summary(tri.geometry())
-        );
-
-        let initial_validation = tri.validate_causality_delaunay();
-        assert!(
-            initial_validation.is_ok(),
-            "Deterministic causal triangle should start causally valid: {initial_validation:?}; {}",
-            deterministic_triangle_debug_summary(tri.geometry())
-        );
-
-        assert!(
-            tri.geometry().faces().any(|face| {
-                tri.face_edge_types(&face)
-                    .is_some_and(|ets| ets.iter().any(|e| matches!(e, EdgeType::Timelike)))
-            }),
-            "Deterministic causal triangle should contain a timelike edge; {}",
-            deterministic_triangle_debug_summary(tri.geometry())
-        );
-
-        let vertex_to_mutate = tri
-            .geometry()
-            .vertices()
-            .next()
-            .expect("Deterministic causal triangle should contain a vertex");
-
-        tri.set_vertex_data(&vertex_to_mutate, Some(3))
-            .expect("Expected valid vertex handle while mutating deterministic triangle");
-
-        let result = tri.validate_causality_delaunay();
-        assert!(
-            result.is_err(),
-            "Explicitly acausal edge should fail causality validation"
-        );
-
-        // Verify the error is a CausalityViolation reporting step_distance > 1.
-        // OpenBoundary topology means step_distance == |Δt|.
-        if let Err(CdtError::CausalityViolation {
-            time_0,
-            time_1,
-            step_distance,
-        }) = result
-        {
-            assert!(
-                step_distance > 1,
-                "CausalityViolation should report step_distance > 1, got step_distance={step_distance} (t0={time_0}, t1={time_1})"
-            );
-            assert_eq!(
-                step_distance,
-                time_0.abs_diff(time_1),
-                "OpenBoundary step_distance must equal |Δt|; got step_distance={step_distance}, |Δt|={}",
-                time_0.abs_diff(time_1)
-            );
-        } else {
-            panic!(
-                "Expected CausalityViolation error, got {result:?}; {}",
-                deterministic_triangle_debug_summary(tri.geometry())
-            );
-        }
-    }
-
-    #[test]
-    fn test_validate_causality_rejects_missing_live_label() {
-        let backend = labeled_triangle_backend([0, 0, 1]);
-        let mut tri = CdtTriangulation::from_labeled_delaunay(backend, 2, 2)
-            .expect("Should preserve labels as foliation");
-        let vertex_to_clear = tri
-            .geometry()
-            .vertices()
-            .next()
-            .expect("Triangle should contain a vertex");
-
-        tri.set_vertex_data(&vertex_to_clear, None)
-            .expect("Expected valid vertex handle while clearing label");
-
-        let result = tri.validate_causality_delaunay();
-
-        assert!(matches!(
-            result,
-            Err(CdtError::ValidationFailed { ref check, ref detail })
-                if check == "causality"
-                    && detail.contains("has no time label in a foliated triangulation")
-        ));
-    }
-
-    #[test]
-    fn test_validate_causality_rejects_all_spacelike_triangle() {
-        let backend = labeled_triangle_backend([0, 0, 0]);
-        let tri = CdtTriangulation::from_labeled_delaunay(backend, 1, 2)
-            .expect("single-slice labels should form foliation bookkeeping");
-
-        let result = tri.validate_causality_delaunay();
-
-        assert!(matches!(
-            result,
-            Err(CdtError::ValidationFailed { ref check, ref detail })
-                if check == "causality"
-                    && detail.contains("invalid CDT triangle")
-                    && detail.contains("spacelike=3")
-                    && detail.contains("timelike=0")
-        ));
-    }
-
-    #[test]
-    fn test_toroidal_causality_violation_reports_circular_step_distance() {
-        // Build a real toroidal CDT with T=10 so wrap-around is non-trivial:
-        // mutating a slice-0 vertex to slice 8 makes its same-slice spacelike
-        // edges read as (raw |Δt|=8, circular distance 2), which exceeds the
-        // adjacency threshold of 1.  The reported step_distance must be the
-        // circular distance, not the raw difference.
-        let mut tri =
-            CdtTriangulation::from_toroidal_cdt(3, 10).expect("build toroidal CDT (3, 10)");
-
-        let slice0_vertex = tri
-            .geometry()
-            .vertices()
-            .find(|vh| tri.geometry().vertex_data_by_key(vh.vertex_key()) == Some(0))
-            .expect("Toroidal CDT should contain slice-0 vertices");
-
-        tri.set_vertex_data(&slice0_vertex, Some(8))
-            .expect("Expected valid vertex handle while mutating label");
-
-        let result = tri.validate_causality_delaunay();
-        match result {
-            Err(CdtError::CausalityViolation {
-                time_0,
-                time_1,
-                step_distance,
-            }) => {
-                let raw = time_0.abs_diff(time_1);
-                let circular = raw.min(10 - raw);
-                assert_eq!(
-                    step_distance, circular,
-                    "Toroidal step_distance must equal circular distance min(|Δt|, T−|Δt|); \
-                     got step_distance={step_distance}, raw |Δt|={raw}, T=10"
-                );
-                assert!(
-                    step_distance > 1,
-                    "Violation must exceed adjacency threshold; got step_distance={step_distance}"
-                );
-                assert!(
-                    step_distance < raw,
-                    "With T=10 and labels {{0,1,8}}, every violation crosses the wrap; \
-                     step_distance={step_distance} should be < raw |Δt|={raw}"
-                );
-            }
-            other => panic!("Expected CausalityViolation on toroidal triangle, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn test_foliation_error_converts_to_cdt_error() {
-        let fol_err = FoliationError::EmptySlice { slice: 3 };
-        let cdt_err: CdtError = fol_err.into();
-
-        match cdt_err {
-            CdtError::Foliation(FoliationError::EmptySlice { slice }) => assert_eq!(slice, 3),
-            other => panic!("Expected Foliation error, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn test_out_of_range_error_conversion() {
-        let fol_err = FoliationError::OutOfRangeVertexLabel {
-            vertex: 2,
-            label: 7,
-            expected_range_end: 2,
-        };
-        let cdt_err: CdtError = fol_err.into();
-
-        match cdt_err {
-            CdtError::Foliation(FoliationError::OutOfRangeVertexLabel {
-                vertex,
-                label,
-                expected_range_end,
-            }) => {
-                assert_eq!(vertex, 2);
-                assert_eq!(label, 7);
-                assert_eq!(expected_range_end, 2);
-            }
-            other => panic!("Expected Foliation error, got {other:?}"),
-        }
-    }
-
-    // =========================================================================
-    // Toroidal CDT tests
-    // =========================================================================
-
-    #[test]
-    fn test_from_toroidal_cdt_basic() {
-        let tri = CdtTriangulation::from_toroidal_cdt(4, 3)
-            .expect("toroidal CDT should build with delaunay v0.7.6");
-
-        // V = N*T = 12, F = 2*N*T = 24, E = 3*N*T = 36, χ = 0.
-        assert_eq!(tri.vertex_count(), 12);
-        assert_eq!(tri.face_count(), 24);
-        assert_eq!(tri.edge_count(), 36);
-        assert_eq!(tri.geometry().euler_characteristic(), 0);
-        assert_eq!(tri.dimension(), 2);
-        assert_eq!(tri.time_slices(), 3);
-        assert!(matches!(tri.metadata().topology, CdtTopology::Toroidal));
-    }
-
-    #[test]
-    fn test_explicit_toroidal_count_validation_rejects_face_mismatch() {
-        let tri = CdtTriangulation::from_toroidal_cdt(4, 3).expect("build toroidal CDT");
-        let result = validate_toroidal_counts(tri.geometry(), 12, 12, 23);
-
-        assert!(matches!(
-            result,
-            Err(CdtError::DelaunayGenerationFailed {
-                vertex_count: 12,
-                coordinate_range: (0.0, 1.0),
-                attempt: 1,
-                ref underlying_error,
-            }) if underlying_error.contains("explicit toroidal builder")
-                && underlying_error.contains("produced 12 vertices and 24 faces")
-                && underlying_error.contains("expected 12 vertices and 23 faces")
-        ));
-    }
-
-    #[test]
-    fn test_from_toroidal_cdt_various_sizes() {
-        for (n, t) in [(3_u32, 3_u32), (4, 3), (5, 4), (6, 5), (8, 4)] {
-            let tri = CdtTriangulation::from_toroidal_cdt(n, t)
-                .unwrap_or_else(|err| panic!("toroidal CDT N={n} T={t} should build: {err}"));
-            let nt = (n as usize) * (t as usize);
-            assert_eq!(tri.vertex_count(), nt);
-            assert_eq!(tri.face_count(), 2 * nt);
-            assert_eq!(tri.edge_count(), 3 * nt);
-            assert_eq!(tri.geometry().euler_characteristic(), 0);
-        }
-    }
-
-    #[test]
-    fn test_from_toroidal_cdt_foliation_per_slice() {
-        let tri = CdtTriangulation::from_toroidal_cdt(5, 4).expect("build toroidal CDT");
-        assert!(tri.has_foliation());
-        assert_eq!(tri.slice_sizes(), &[5, 5, 5, 5]);
-        for t in 0..4 {
-            assert_eq!(
-                tri.vertices_at_time(t).len(),
-                5,
-                "slice {t} should contain N=5 vertices"
-            );
-        }
-    }
-
-    #[test]
-    fn test_from_toroidal_cdt_validate_passes() {
-        let tri = CdtTriangulation::from_toroidal_cdt(4, 3).expect("build toroidal CDT");
-        assert!(tri.validate_topology().is_ok());
-        assert!(tri.validate_foliation().is_ok());
-        assert!(tri.validate_causality().is_ok());
-    }
-
-    #[test]
-    fn test_from_toroidal_cdt_each_slice_is_closed_s1() {
-        // The toroidal validate_foliation extension already enforces that
-        // each slice is a closed S¹; building a torus and validating again
-        // exercises that path explicitly.
-        let tri = CdtTriangulation::from_toroidal_cdt(6, 4).expect("build toroidal CDT");
-        tri.validate_foliation()
-            .expect("explicit toroidal CDT must satisfy closed-S¹ per-slice invariant");
-    }
-
-    #[test]
-    fn test_from_toroidal_cdt_invalid_params() {
-        // Too few vertices per slice (N < 3 — no proper cycle).
-        assert!(CdtTriangulation::from_toroidal_cdt(2, 3).is_err());
-        // Too few slices (T < 3 — wrap-around makes the mesh non-manifold).
-        assert!(CdtTriangulation::from_toroidal_cdt(4, 1).is_err());
-        assert!(CdtTriangulation::from_toroidal_cdt(4, 2).is_err());
-    }
-
-    #[test]
-    fn test_from_toroidal_cdt_rejects_vertex_count_overflow() {
-        let result = CdtTriangulation::from_toroidal_cdt(u32::MAX, 3);
-
-        assert!(matches!(
-            result,
-            Err(CdtError::InvalidGenerationParameters {
-                ref issue,
-                ref provided_value,
-                ref expected_range,
-            }) if issue == "Vertex count overflow"
-                && provided_value == "4294967295 × 3"
-                && expected_range == "product ≤ u32::MAX"
-        ));
-    }
-
-    #[test]
-    fn test_toroidal_cell_classification_uses_temporal_wrap() {
-        let tri = CdtTriangulation::from_toroidal_cdt(4, 3).expect("build toroidal CDT");
-        let mut saw_wrap_up = false;
-        let mut saw_wrap_down = false;
-        let mut saw_wrap_timelike_edge = false;
-
-        for face in tri.geometry().faces() {
-            let vertices = tri
-                .geometry()
-                .face_vertices(&face)
-                .expect("toroidal face vertices should resolve");
-            let labels: Vec<_> = vertices
-                .iter()
-                .map(|vh| {
-                    tri.geometry()
-                        .vertex_data_by_key(vh.vertex_key())
-                        .expect("toroidal vertices are labeled")
-                })
-                .collect();
-
-            if labels.contains(&0) && labels.contains(&2) {
-                let cell_type = tri
-                    .cell_type(&face)
-                    .expect("wrap-around toroidal face should classify");
-                let edge_types = tri
-                    .face_edge_types(&face)
-                    .expect("wrap-around toroidal face should expose edge types");
-                saw_wrap_timelike_edge |= edge_types
-                    .iter()
-                    .any(|edge_type| matches!(edge_type, EdgeType::Timelike));
-
-                let zero_count = labels.iter().filter(|&&label| label == 0).count();
-                let two_count = labels.iter().filter(|&&label| label == 2).count();
-                let is_wrap_up = zero_count == 1 && two_count == 2;
-                let is_wrap_down = zero_count == 2 && two_count == 1;
-
-                if is_wrap_up {
-                    assert_eq!(cell_type, CellType::Up);
-                }
-                if is_wrap_down {
-                    assert_eq!(cell_type, CellType::Down);
-                }
-
-                saw_wrap_up |= is_wrap_up;
-                saw_wrap_down |= is_wrap_down;
-            }
-        }
-
-        assert!(saw_wrap_up, "expected an Up cell across the temporal wrap");
-        assert!(
-            saw_wrap_down,
-            "expected a Down cell across the temporal wrap"
-        );
-        assert!(
-            saw_wrap_timelike_edge,
-            "expected a timelike edge across the temporal wrap"
-        );
-    }
-
     #[test]
     fn test_validate_topology_open_boundary() {
         let tri = CdtTriangulation::from_seeded_points(5, 1, 2, 53).expect("create triangulation");
@@ -4701,29 +1237,6 @@ mod tests {
         ));
         assert_eq!(tri.time_slices(), 3);
         assert!(tri.validate_topology().is_ok());
-    }
-
-    #[test]
-    fn test_assign_foliation_by_y_rejects_invalid_toroidal_slice_count() {
-        let mut tri = CdtTriangulation::from_toroidal_cdt(4, 3).expect("build toroidal CDT");
-        let initial_slice_sizes = tri.slice_sizes().to_vec();
-
-        let result = tri.assign_foliation_by_y(2);
-        assert!(matches!(
-            result,
-            Err(CdtError::InvalidTriangulationMetadata {
-                ref field,
-                ref topology,
-                ref provided_value,
-                ref expected,
-            }) if field == "timeslices"
-                && topology == "toroidal"
-                && provided_value == "2"
-                && expected == "≥ 3"
-        ));
-        assert_eq!(tri.time_slices(), 3);
-        assert_eq!(tri.slice_sizes(), initial_slice_sizes.as_slice());
-        assert!(tri.has_foliation());
     }
 
     #[test]
